% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{acond}
\alias{acond}
\alias{aprop}
\title{Transform a frequency table into relative frequencies relative to a margin.}
\usage{
acond(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")

aprop(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")
}
\description{
Transform a frequency table into relative frequencies relative to a margin.
}

