% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{add_faces}
\alias{add_faces}
\title{Add faces to an array of data frames}
\usage{
add_faces(arr, label = "All", ...)
}
\arguments{
\item{arr}{an array of data frames}

\item{label}{character string to label faces, default "All"}
}
\description{
Add union of rows in elements of an array of data frames to marginal
faces of the array
}
