% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{anon}
\alias{anon}
\title{Remove names}
\usage{
anon(x)
}
\arguments{
\item{x}{object that may have names}
}
\value{
x without names
}
\description{
Anonymize a vector or list
}
\details{
Useful is a list has names but will be passed as the argument list to \code{\link{do.call}} 
if the names are not intended to be parameter names.
}
\seealso{
\code{\link[spida2]{name}}

\code{\link[spida2]{auto}}
}
