% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{append<-}
\alias{append<-}
\title{Assignment function to append to a list}
\usage{
append(x, after = length(x)) <- value
}
\arguments{
\item{x}{list to which to append}

\item{after}{where, default: end}
}
\description{
Assignment function to append to a list
}
\examples{
zzz <- list()
append(zzz) <- 'a'
append(zzz) <- 'b'
append(zzz,1) <- list('d','e')
append(zzz) <- list('f','g')
zzz
}
