% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{Apply}
\alias{Apply}
\title{Version of lapply that can return an array of lists}
\usage{
Apply(X, fun, ...)
}
\arguments{
\item{X}{list of array of lists, as in \code{\link{lapply}}}

\item{fun}{function to be applied to each element of X}

\item{...}{same as \code{\link{lapply}}}
}
\description{
Returns an array of lists similar to its input
}
