% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{approx.Z}
\alias{approx.Z}
\title{Approximation for semi-parametric splines}
\usage{
approx.Z(Z, oldtimes, newtimes)
}
\arguments{
\item{Z}{}

\item{oldtimes}{}

\item{newtimes}{}
}
\description{
Approximation for semi-parametric splines
}

