% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{as.data.frame.wald}
\alias{as.data.frame.wald}
\title{Transform output of a Wald test into a data frame}
\usage{
\method{as.data.frame}{wald}(obj, se = 2, digits = 3, sep = "",
  which = 1)
}
\arguments{
\item{obj}{a 'wald' object}

\item{se}{a vector with the multiples of standard error used to generate lower and upper limits. 'names(se)'
is appended to 'L' and 'U' to label the variables.}

\item{which}{selects elements of 'obj' to turn to a data.frame.}
}
\value{
A data frame with estimated coefficient, standard error, and, optionally, upper and lower limits and
        the variables included the 'data' element of 'obj' if present.
        If \code{length(which) > 1}, the returned object is a list of data frames.
}
\description{
The generic method is included in the possibly false hope that it will ensure that the method
in this package is used.
}
\examples{
# coming soon!
}
