% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{assn}
\alias{assn}
\title{Assign in a pipeline}
\usage{
assn(value, x, pos = 0, ...)
}
\arguments{
\item{value}{to be assigned}

\item{x}{name of the object to be assigned}

\item{pos}{where the object is to be assigned. Default is \code{pos=0} which
saves in the global environment}

\item{...}{other arguments are passed to \code{assign}}
}
\value{
The formula is evaluated using \code{model.frame} and the
result is used as the argument of \code{order} which, in turn is used to
order the data frame. The ordered data frame is returned.
}
\description{
Same as \code{assign} with the first two arguments reversed and the default
position referrring to the global environment so it can be used
interactively in a pipeline
}
\examples{
\dontrun{
# Using the Prestige data set in package:car and the pipeline
# function in package:magrittr, we can sort the data frame and save the sorted
# version as "Prestige.ordered" with:
require(magrittr)
require(car)
Prestige  \%>\% sortdf(~type+income)  \%>\% assn("Prestige.ordered")  \%>\%  dim
}
}

