% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{atotal}
\alias{atotal}
\title{Add all marginal sums to an array}
\usage{
atotal(arr, FUN = sum, label = "Total", ...)
}
\arguments{
\item{arr}{array for which margins need to be added}

\item{FUN}{function to compute margins, default: sum}

\item{label}{for margin, default: 'Total'}

\item{\dots}{additional arguments to 'FUN'}
}
\value{
an array with dimension dim(arr) + 1
}
\description{
atotal adds by default a border of sums to an array. The function FUN may be used instead of 'sum'. Additional
arguments to FUN can also be given.
}
\author{
Georges Monette
}
