% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{auto}
\alias{auto}
\alias{autonym}
\title{Creates an autonym}
\usage{
auto(x, force = FALSE)

autonym(x, force = FALSE)
}
\arguments{
\item{x}{object that can be used to name itself.}

\item{force}{(default FALSE) should x replace its name if it already has one.}
}
\value{
x with names equal to x
}
\description{
Uses x to name x. Useful in when applying \code{\link{lapply}} to a character vector.
Should probably be called 'autonym'. Superseded by \code{\link{name}}.
}
\seealso{
\code{\link{anon}}
}
