% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{basis}
\alias{basis}
\title{Orthogonal basis for column space of a matrix}
\usage{
basis(X, coef = FALSE)
}
\arguments{
\item{X}{}

\item{coef}{}
}
\description{
Orthogonal basis for column space of a matrix
}

