% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{bm}
\alias{bm}
\title{Another benchmark}
\usage{
bm(n = 1)
}
\arguments{
\item{n}{number of replications, default 1}
}
\description{
This benchmark is designed to show results for specific funtions.
}
\details{
For example. qr is not parallelized with openblas-open mp
Timings are compared with those on an  AMD Ryzen 9 3950X 16-Core Processor
with 16 cores and not hyperthreading.
}
