% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{bm25}
\alias{bm25}
\title{Simon Urbanek's Benchmark 2.5}
\usage{
bm25()
}
\description{
Use the total time for all 15 tests to decide whether you would benefit
from running time-consuming tasks on a faster server.
}
\seealso{
\code{\link[=bm25f]{bm25f()}} for a benchmark saving results in a file
}
