% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm25.R
\name{bm25f}
\alias{bm25f}
\title{Simon Urbanek's Benchmark 2.5}
\usage{
bm25f(note = "none specified")
}
\description{
Use the total time for all 15 tests to decide whether you would benefit
from running time-consuming tasks on a faster server.
}
\details{
Exports information to a file
}
\seealso{
\code{\link[=bm25]{bm25()}} for a benchmark without saving results in a file
}
