% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brace.R
\name{brace}
\alias{brace}
\title{Draw a brace}
\usage{
brace(x1 = 0, y1 = 0, x2 = 0, y2 = 1, right = TRUE, rad = 0.2)
}
\arguments{
\item{x1, }{y1, x1, y2 endpoints of brace}

\item{right}{(default TRUE) direction in which brace points}

\item{rad}{(default 0.2) radius of curl of brace}
}
\description{
Draw a brace
}
\examples{
plot(c(0,10), c(0,10), type = 'n')
lines(brace(0, 0, 0, 5), lwd = 2)
text(1.2,2.5,"size", adj = 0)
plot(c(0,10),c(0,1))
brace(5,0,5,1) \%>\% lines
lapply(1:10, function(x) {
   lines(brace(x,0,x,.1*x,F))
   lines(brace(x,0,x-.2*x,.1*x,F),col = 'red')
})
}
\author{
Georges Monette
}
\concept{lattice}
