% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{brewer.pal.show}
\alias{brewer.pal.show}
\title{Display RColorBrewer palette}
\usage{
brewer.pal.show(
  n = NULL,
  type = "all",
  select = NULL,
  exact.n = TRUE,
  colorblindFriendly = FALSE
)
}
\description{
Function name that follows usual conventions and is
easier to remember.
}
