% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{C_}
\alias{C_}
\title{Contrasts to make intercept a weighted average}
\usage{
C_(x, weights = rep(1, length(levels(x))))
}
\arguments{
\item{x}{a factor}

\item{weights}{a vector fo relative weights applied to
levels of \code{x}}
}
\description{
A generalization of treatment contrasts that make
the intercept a weighted average of factor levels
(instead of placing all the weight on the reference level)
}
\examples{
x <- factor(rep(letters[1:4],4))
y <- 1:length(x)
lm(y ~ x)
xx <- C_(x)
lm(y ~ xx)
}
\seealso{
\code{\link[=contr.nhelmert]{contr.nhelmert()}} for weighted average
}
