% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R
\name{ccor}
\alias{ccor}
\title{Canonical correlation on uncentered matrices}
\usage{
ccor(x, y)
}
\arguments{
\item{x, y}{two matrices with the same number of rows}
}
\description{
The vector of canonical correlations of uncentered
matrices helps to identify whether the column
spaces of two matrices are identical, or whether one
is a subspace of the other, or whether the spaces
are orthogonal.
}
