% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{cell}
\alias{cell}
\alias{cell.wald}
\alias{cell.default}
\title{Confidence ellipse}
\usage{
cell(x, ...)

\method{cell}{wald}(
  obj,
  which.coef = 1:2,
  levels = 0.95,
  Scheffe = FALSE,
  dfn = 2,
  center.pch = 19,
  center.cex = 1.5,
  segments = 51,
  xlab,
  ylab,
  las = par("las"),
  col = palette()[2],
  lwd = 2,
  lty = 1,
  add = FALSE,
  ...
)

\method{cell}{default}(
  model,
  which.coef,
  levels = 0.95,
  Scheffe = FALSE,
  dfn = 2,
  center.pch = 19,
  center.cex = 1.5,
  segments = 51,
  xlab,
  ylab,
  las = par("las"),
  col = palette()[2],
  lwd = 2,
  lty = 1,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{fitted object}

\item{\dots}{other arguments passed to method}

\item{which.coef}{(default 1:2) which parameters to estimate}

\item{levels}{(default 0.95)}

\item{dfn}{(default 2) degrees of freedom for Sheffe multiple test and confidence limits}

\item{Sheffe}{logical (default FALSE)}
}
\description{
Confidence ellipse for two parameters of a fitted object
}
\section{Methods (by class)}{
\itemize{
\item \code{cell(wald)}: method for wald object

\item \code{cell(default)}: default method

}}
