% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{change}
\alias{change}
\title{Replace elements of x with correspondingly named elements of ll}
\usage{
change(x, ...)

change(x, ...)
}
\arguments{
\item{x}{list or vector some of whose elements will be replaced or appended to}

\item{...}{elements for replacement or appending}

\item{ll}{FIXME}
}
\description{
Replace elements of x with correspondingly named elements of ll

Replace elements of x with correspondingly named elements of ll
}
