% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{coef.wald}
\alias{coef.wald}
\title{Extract coefficients from Wald object}
\usage{
\method{coef}{wald}(obj, se = FALSE)

\method{coef}{wald}(obj, se = FALSE)
}
\arguments{
\item{obj}{wald object}

\item{se}{(default: FALSE) include standard errors}
}
\value{
coefficients and, optionally, standard errors

coefficients and, optionally, standard errors
}
\description{
Extract coefficients from Wald object

Extract coefficients from Wald object
}
\examples{
# coming soon
# coming soon
}
\author{
GM
}
