% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R, R/colours2.R
\name{col2hex}
\alias{col2hex}
\title{Color to hex}
\usage{
col2hex(cname)

col2hex(cname)
}
\arguments{
\item{cname}{color name

Borrowed from 'gplots'}
}
\description{
Color to hex

Color to hex
}
