% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R, R/colours2.R
\name{cols}
\alias{cols}
\title{Colour collection}
\usage{
cols(pattern = NULL)

cols(pattern = NULL)
}
\arguments{
\item{pattern}{return a list of colour palettes whose names match \code{pattern}.
If only one palette matches, it is returned as a vector.  If
NULL (default), return a list with all colour palettes.
The pattern 'cb' returns the Okabe-Ito palette
designed to accommodate predominant forms of colour blindness.
See \href{https://easystats.github.io/see/reference/scale_color_okabeito.html}{Okabe-Ito Color Palette}.}
}
\description{
A collection of colours found in various places

A collection of colours found in various places
}
\examples{
names(cols())
pal(cols()$mycols)
pal(cols()$tol8)
lapply(names(cols()),function(x) pal(cols()[[x]], main = x))
names(cols())
pal(cols()$mycols)
pal(cols()$tol8)
lapply(names(cols()),function(x) pal(cols()[[x]], main = x))
}
