% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{contr.nhelmert}
\alias{contr.nhelmert}
\title{Normalized Helmert contrasts for parsimonious random effects models}
\usage{
contr.nhelmert(n, ...)
}
\arguments{
\item{n}{a vector of levels for a factor or the number of levels.}
}
\description{
\itemize{
\item \href{https://stats.idre.ucla.edu/r/library/r-library-contrast-coding-systems-for-categorical-variables/}{R Library Contrast Coding Systems for categorical variables}
}
}
\examples{
zf <- factor(c("A","B","B","C"))
contrasts(zf) <- contr.nhelmert(3)
contrasts(zf)
library(nlme)
set.seed(123)
n_within <- 10
n_schools <- 20
n_treatments <- 3
dd <- expand.grid(id = 1:n_within, schoolid = 1:n_schools)
dd <- within(
   dd,
   {
      method <- factor(sample(LETTERS[1:n_treatments], nrow(dd), TRUE))
      methodn <- C(method, contr.nhelmert)
      methodmat <- contrasts(methodn)[methodn,]
      grade <- 10 + seq_len(n_treatments)[method] + rnorm(n_schools)[schoolid] + rnorm(nrow(dd))
   })
fitfull <- lme(grade ~ method, dd, random = ~ 1 + method | schoolid,
 control = list(returnObject = T))     
summary(fitfull)
getG(fitfull) \%>\% svd(nu=0,nv=0)
fitfulln <- lme(grade ~ method, dd, random = ~ 1 + methodn | schoolid,
 control = list(returnObject = T))     
summary(fitfulln)
getG(fitfulln)
getG(fitfulln) \%>\% svd(nu=0,nv=0)
fitpars1 <- lme(grade ~ method, dd, 
   random = list( schoolid = pdBlocked(list(pdDiag(~1), pdIdent(~ methodn - 1))) ))
fitpars2 <- lme(grade ~ method, dd, 
   random = list( schoolid = pdBlocked(list(pdDiag(~1), pdIdent(~ methodmat - 1))) ))
summary(fitpars1)
summary(fitpars2)
ranef(fitpars2) 
}
