% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capply.R
\name{cvar}
\alias{cvar}
\title{Create a contextual variable for regression}
\usage{
cvar(x, id, ...)
}
\arguments{
\item{x}{variable to be centered or residualized within groups. If x is a
factor, cvar and dvar return matrices whose columns are named consistently
with the names of coding variables for factors.}

\item{id}{identifies clusters}
}
\description{
cvar and dvar are designed to be used in regression formulas to
create a contextual mean of a cluster-varying variable and
a 'centered-within-groups' version.
}
\examples{
\dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
}
}

