% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{debugged}
\alias{debugged}
\title{List debugged functions}
\usage{
debugged(environments = search(), all = FALSE)
}
\arguments{
\item{environments}{default: search()}

\item{all}{list all functions including those not debugged, default: FALSE}
}
\value{
a data frame listing functions and whether they are currently debugged
}
\description{
List debugged functions
}
