% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{DiffI}
\alias{DiffI}
\title{Contextual diff with intrapolation}
\usage{
DiffI(xx, ...)
}
\arguments{
\item{x}{vector to be lagged}

\item{id}{clustering variable}

\item{idx}{time variable}

\item{lag}{(default 1)}

\item{at}{alternative to 'lag': evaluate x at a particular value of 'idx'}

\item{check}{(default TRUE) check for unique values of idx within each cluster}
}
\value{
return the value of  x minus x in the position 'idx - lag' in the same cluster 
        using intrapolation
}
\description{
Diffs within clusters based on a time vector with possibly unequal intervals
}
\seealso{
\code{\link{LagI}}, \code{\link{DiffI}}, \code{\link{Diff}},
}
