% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{dropLastTotal}
\alias{dropLastTotal}
\title{Drop last elements of an array if it is a "Total"}
\usage{
dropLastTotal(mat, names_to_drop = "Total", drop = FALSE)
}
\arguments{
\item{mat}{a matrix, array or table}

\item{names_to_drop}{(default "Total")}

\item{drop}{(default FALSE) should one-element dimension be dropped as dimensions}
}
\description{
Used to drop "Total" rows and columns after using \code{\link{tab}}. Synonyms for
legacy: \code{Tab} and \code{pab}.
}
\seealso{
tab, Tab, dropLast
}
