% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{esac}
\alias{esac}
\title{Sequential ifelse with paired arguments}
\usage{
esac(..., OTHER = "NONE SELECTED")
}
\arguments{
\item{...}{a sequence of alternating arguments with each pair consisting
of a vector logical argument followed by a vector of values to be returned in
the positions in which the logical argument is TRUE. Each pair
corresponds to the first two arguments of a \code{\link{ifelse}}.}

\item{OTHER}{value returned if no logical vector evaluates to TRUE in 
a position.  Can be modified directly or the same result achieved
by ending the pairs of arguments with: TRUE, other_value.}
}
\value{
a vector consisting of the value vector corresponding to the first
       logical vector that evaluates to TRUE in a position.
}
\description{
Equivalent of nested ifelse with alternating conditions and values
}
