% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{expandModelData}
\alias{expandModelData}
\title{Add rows of predictor variable values to model data frame}
\usage{
expandModelData(model, add)

expandModelData(model, add)
}
\arguments{
\item{add}{a data frame containing values to be added.}

\item{mod}{a model from which a model data frame can be
extracted with \code{\link{getModelData}}.}
}
\value{
a data frame with concatenating the rows of the
       model data frame and the rows of 'add' together
       with an additional variable '.source' with the
       value 'model' or 'add' showing the provenance of the row.

a data frame with concatenating the rows of the
       model data frame and the rows of 'add' together
       with an additional variable '.source' with the
       value 'model' or 'add' showing the provenance of the row.
}
\description{
Generates additional rows of a predictor data frame
so that factors retain the levels of the factors
in the model data frame,  provided no new values
are introduced.

Generates additional rows of a predictor data frame
so that factors retain the levels of the factors
in the model data frame,  provided no new values
are introduced.
}
