% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{f2R}
\alias{f2R}
\title{Turns a factor of a PD matrix to a right-triangular factor}
\usage{
f2R(A)
}
\arguments{
\item{x}{a factor of a positive definite matrix $V$}
}
\value{
a right-triangular factor of the same $V$.
}
\description{
This function takes a factor of a positive-definite matrix ($V = A'A$ ) and
returns the equivalent right-triangular factor ($V = R'R$).
}
\seealso{
\code{f2R}
}
