% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpanel.R
\name{fillin}
\alias{fillin}
\title{Fill in values}
\usage{
fillin(data, form, n = 200, xpd = 1)
}
\arguments{
\item{data}{data frame with values of x that need filling in}

\item{form}{formula idenfying variable x to fill in and grouping variables, g1, g2, etc.
using syntax: \code{~ x + g1 + g2} (the variable to fill in comes first)}

\item{n}{number of additional points over range of predictor (default 200)}

\item{xpd}{expansion beyond range of predictor (default 1.0, i.e. no expansion)}
}
\description{
Fills in values in gaps between observed predictor values to help produce
a smooth graph of predicted values versus predictor values.
}
