% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbind.R
\name{fix_xy}
\alias{fix_xy}
\title{Fix .x and .y variables in a merged data frame}
\usage{
fix_xy(data)

fix_xy(data)
}
\arguments{
\item{data}{a data frame that is the result of a merge operation, possibly
with .x and .y variables

Note that factors will merge into a common variable only if they have the same
set of values. To merge factors with different sets of values, consider
converting them to character strings.}
}
\description{
Function to combine consistent .x and .y variable names after a merge

Some pairs will be identical and they should become one variable without the suffix.
Some will be compatible in the sense that non-missing values are identical and
values missing in one but not the other can be merged
Some will be incompatible and they should remain in their .x and .y version.
}
\details{
Some pairs will be identical and they should become one variable without the suffix.
Some will be compatible in the sense that non-missing values are identical and
values missing in one but not the other can be merged
Some will be incompatible and they should remain in their .x and .y version.

The idea is that after applying this function, any .x and .y variable pairs
are, in fact, distinct variables that need to be looked at before being used
in any analyses.

The idea is that after applying this function, any .x and .y variable pairs
are, in fact, distinct variables that need to be looked at before being used
in any analyses.
}
\examples{
da <- data.frame(
  id = 1:5, 
  g = c(1,2,NA,4,5), 
  h = factor(c('a','b','c','d','e')),
  hh = c('a','b','c','d','e'),
  m = factor(c('M',NA,NA,'F','M'))  # factor with same set of values
)
db <- data.frame(
  id = c(1:3,6), 
  g = c(1,2,NA,6), 
  h = factor(c('a','b','c','f')),  # factor with different sets of values
  hh = c('a','b','c','f'),
  k = c(1:3,6),
  m = factor(c('M','M',NA,'F'))
)
da  
db  
dm <- dplyr::full_join(da, db, by = 'id' )  
dm
fix_xy(dm)
}
