% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{td}
\alias{td}
\title{Set lattice parameters for multiple groups}
\usage{
td(...)
}
\arguments{
\item{\dots}{FIXME
parameter: e.g. \code{plot.symbol=list(cex=2,col='red')}. Particular useful
for the cex, col, lty, lwd, alpha, pch parameters in plot.line and
plot.symbol.}

\item{n}{in 'gd' specifies the number of distinct colours to generate to
distinguish groups. 'gd' uses 'latticeExtra' to set defaults for a
ggplot2-like appearance. Default is n = 4}

\item{new}{If new = TRUE, open a new window, otherwise modify the existing
active window, if there is one.}

\item{col}{FIXME}

\item{lty}{FIXME}

\item{lwd}{FIXME}

\item{pch}{FIXME}

\item{cex}{FIXME
for each level of the groups variable}

\item{font}{FIXME}

\item{fill}{FIXME}

\item{col.line}{FIXME
not given}

\item{col.symbol}{FIXME
'groups' not given}

\item{alpha}{FIXME}

\item{alpha.line}{FIXME}

\item{alpha.symbol}{graphical parameters for superpose.line and
superpose.symbol}

\item{len}{extend the length of parameters by recycling to length 'len'}

\item{long}{if TRUE generate a default combination of col, lty and pch with
length 42.}

\item{record}{If TRUE, set history to 'recording'. Caution: this can use a
lot of memory in .GlobalEnv.  Consider adding selected graphs to memory with
the 'Insert' key instead.}

\item{basecol}{FIXME}

\item{colsets}{FIXME}
}
\description{
Easier alternative to using trellis.par.set and trellis.par.get to change lattice colors, line types, etc.
}
\details{
Designed to easily set lattice parameters for multiple groups. Setting
parameters before calling the lattice function allows parameters to be used
consistently in the group key.

'td' calls 'trellis.device' and sets graphical parameters for
'superpose.line' and 'superpose.symbol'. 'td' also initializes a new trellis
device with a white background if new = TRUE.

'gd' is similar to 'td' except that it uses a theme that resembles that of 'ggplot'
}
\section{Functions}{
\itemize{
\item \code{td}: without ggplot2-like theme
}}

\examples{
td( lty = 1:7)   # sets line types for 7 groups
gd(7)            # sets line types for 7 groups using colors from RColorBrewer
td( plot.symbol = list(col = 'red', pch = 17))
gd_(col='blue')  # set main color to 'blue'
}
\author{
Georges Monette
}
\concept{lattice}
