% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{getData}
\alias{getData}
\alias{getData.lmer}
\alias{getData.lme}
\alias{getData.gls}
\alias{getData.lm}
\title{Extract the model data frame for various fitting methods}
\usage{
getData(x, ...)

\method{getData}{lmer}(x, ...)

\method{getData}{lme}(x, ...)

\method{getData}{gls}(x, ...)

\method{getData}{lm}(x, ...)
}
\arguments{
\item{fit}{a fitted object}
}
\description{
getData is to lme what model.frame is to lm
getData is implemented as a method in nlme
so we just add a method for 'lm' and other objects
}
\section{Methods (by class)}{
\itemize{
\item \code{getData(lmer)}: method for lmer objects

\item \code{getData(lme)}: method for lme objects

\item \code{getData(gls)}: method for gls objects

\item \code{getData(lm)}: method for lm objects

}}
