% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{getex}
\alias{getex}
\title{Extract or remove a substring}
\usage{
getex(x, pattern, ...)
}
\arguments{
\item{x}{object from which strings are extracted. Will be coerced to a
character by \code{\link{regexpr}}.}

\item{pattern}{a regular expression whose matches in \code{x} are returned.}

\item{...}{other arguments passed to \code{\link{regexpr}}.}
}
\value{
\code{getex} returns the substrings matched by \code{pattern} in \code{x}. 
        \code{getex_} returns the portion of \code{x} that is not matched.
}
\description{
Extract of remove a substring matching a regular expression.
}
\details{
\code{getex} and \code{getex_} can work as a pair to gradually extract information
from a string.
}
\examples{
library(magrittr)
# The following could be done with 'strsplit' but 'getex' can handle messier cases
progs <- c('SC BSc Honours Biology', 'Arts BA Ordinary Economics')
faculty <- progs  \%>\%  getex('^[^ ]+') \%>\% toupper
progs <- progs  \%>\% getex_('^[^ ]+ ')
degree <- progs \%>\% getex('^[^ ]+')
}
