% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{getFix}
\alias{getFix}
\alias{getFix.MCMCglmm}
\alias{getFix.default}
\alias{getFix.glm}
\alias{getFix.glmer}
\alias{getFix.gls}
\alias{getFix.lm}
\alias{getFix.lme}
\alias{getFix.lmer}
\alias{getFix.mer}
\alias{getFix.mipo}
\alias{getFix.multinom}
\alias{getFix.zeroinfl}
\title{Get information on fixed effects from a model object}
\usage{
getFix(fit, ...)

\method{getFix}{multinom}(fit, ...)

\method{getFix}{lm}(fit, ...)

\method{getFix}{glm}(fit, ...)

\method{getFix}{lme}(fit, ...)

\method{getFix}{gls}(fit, ...)

\method{getFix}{lmer}(fit, ...)

\method{getFix}{glmer}(fit, ...)

\method{getFix}{mer}(fit, ...)

\method{getFix}{zeroinfl}(fit, ...)

\method{getFix}{mipo}(fit, ...)

\method{getFix}{MCMCglmm}(fit, ...)

\method{getFix}{default}(fit, ...)
}
\arguments{
\item{fit}{A fitted model object}

\item{\dots}{Other arguments [unused]}
}
\value{
Returns a list with the following components:
     \itemize{
          \item{fixed}{Fixed effect parameter estimates}
          \item{vcov}{Covariance matrix of the parameters}
          \item{df}{denominator degrees of freedom for each effect}
          }
}
\description{
\code{getFix} extracts coefficients, covariance matrix and degrees of
freedom from model objects. Its main purpose is to extract information need
by the \code{wald} function. To extend the wald function to a new class of
objects, it is merely necessary to write a method for \code{getFix}.
}
\details{
Extending the \code{\link{wald}} function to a new class of objects only
requires a \code{getFix} method for the new class.
}
\section{Methods (by class)}{
\itemize{
\item \code{multinom}: method for multinom objects in package nnet

\item \code{lm}: method for lm objects

\item \code{glm}: method for glm objects

\item \code{lme}: method for lme objects in the nlme package

\item \code{gls}: method for gls objects in the nlme package

\item \code{lmer}: method for lmer objects in the lme4 package

\item \code{glmer}: method for glmer objects in the lme4 package

\item \code{mer}: method for mer objects in the lme4 package

\item \code{zeroinfl}: method for zeroinfl objects in the pscl?? package

\item \code{mipo}: method for mipo objects in the mice package

\item \code{MCMCglmm}: method for MCMCglmm objects in the MCMCglmm package

\item \code{default}: print message if getFix id used for a class for which a method has not been written
}}
\examples{
library(nlme)
fit <- lme( mathach ~ (ses + I(ses^2)) * Sex, hs, random = ~ 1 + ses| school)
getFix(fit)

data(Prestige, package="car")
mod.prestige <- lm(prestige ~ education + income, data=Prestige)
getFix(mod.prestige)

}
\author{
Georges Monette
}
\seealso{
\code{\link{wald}}
}

