% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{getModelData}
\alias{getModelData}
\title{Get the data used to fit a model}
\usage{
getModelData(model)

getModelData(model)
}
\arguments{
\item{model}{a model to which the function \code{\link{model.frame}} 
can be applied along with a number of other basic functions and
methods.}
}
\value{
a data frame with the variables needed to fit 'model'.

a data frame with the variables needed to fit 'model'.
}
\description{
Returns a data frame with the variables used to fit a model.
In contrast with \code{\link{model.frame}}, \code{getModelData}
returns the variables as they appear in the original data frame
used to fit the model.

Returns a data frame with the variables used to fit a model.
In contrast with \code{\link{model.frame}}, \code{getModelData}
returns the variables as they appear in the original data frame
used to fit the model.
}
