% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS.R
\name{getS}
\alias{getS}
\title{Extract sample and sampling parameters from Bayesian posterior}
\usage{
getS(x, ...)
}
\arguments{
\item{x}{a fitted Bayesian regression object, for example,
using \code{rstan}.}
}
\description{
Extract sample and sampling parameters from Bayesian posterior
}
