% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{getX}
\alias{getX}
\title{Design matrix for a fit possibly on a new data frame}
\usage{
getX(fit, data = getModelData(fit))

getX(fit, data = getModelData(fit))
}
\arguments{
\item{fit}{a fitted object with formula method}

\item{data}{(default NULL) a data frame on which to evaluate the design matrix}
}
\value{
a design matrix

a design matrix
}
\description{
This function return the X matrix for a fit possibly based on a different data frame than the model.
It performs a function very close to that of \code{\link{model.matrix}} except that \code{\link{model.matrix}}
expects the variable for the LHS of the formula to be in the data set, ostensibly in order to remove rows
for which the LHS variable(s) are NA. In addition, \code{getX} attaches the argument data set as an attribute.

This function return the X matrix for a fit possibly based on a different data frame than the model.
It performs a function very close to that of \code{\link{model.matrix}} except that \code{\link{model.matrix}}
expects the variable for the LHS of the formula to be in the data set, ostensibly in order to remove rows
for which the LHS variable(s) are NA. In addition, \code{getX} attaches the argument data set as an attribute.
}
\details{
Extending \code{getX} to new classes merely requires a \code{\link{getData}} method. The \code{\link{formula}}
method is also used but usually already exitsts.

Extending \code{getX} to new classes merely requires a \code{\link{getData}} method. The \code{\link{formula}}
method is also used but usually already exitsts.
}
