% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gicc.R
\name{gicc}
\alias{gicc}
\title{Generalized intraclass correlation coefficient for multilevel data}
\usage{
gicc(x, by, method, ...)
}
\arguments{
\item{x}{a data frame, factor, character variable or a numerical variable.}

\item{by}{if \code{x} is a data frame, \code{by} can be a formula (e.g.
\code{~id}) evaluated in \code{x}. Otherwise, \code{by} is a variable
defining clusters.}

\item{method}{a character string indicating whether to work with the rank of
the raw variable, \code{x}, stripped of missing values, with the raw
variable itself, or with \code{is.na(x)}. Can be one of "rank", raw" or
"na".}

\item{\dots}{-- not used}
}
\value{
a measure of relative variability within
clusters so that 1
represents no variability and 0 perfect balance.
}
\description{
\code{gicc} facilitates the identification
of between-subject variables
versus balanced within-subject variables with
hierarchical data. Methods
handle factors, character variables and
numerical variables. For categorical
variables, Goodman-Kruskal's tau is returned and
for numerical variables,
the simple ICC with (variance between) / (variance between + variance
within) applied to the rank (by default) of the variable. In either case, a
value of 1 signifies a variable that is constant within clusters and a value
of 0, a variable that is perfectly balanced within clusters.
}
\examples{
  gicc( hs, ~ school)
  gicc( hs, ~ school, method = 'raw')
  gicc( hs, ~ school/sex)
}
\references{
see p. 25 of Agresti (1990) Categorical Data Analysis, Wiley.
}

