% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less.R
\name{\%and\%}
\alias{\%and\%}
\title{intersect as a binary operator}
\usage{
x \%and\% y
}
\arguments{
\item{x, y}{vectors of the same mode containing sequences of items}
}
\value{
a vector of the same mode containing the intersection of \code{x} less \code{y}
}
\description{
The intersection of x and y
}

