% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less.R
\name{\%less\%}
\alias{\%less\%}
\title{setdiff as a binary operator}
\usage{
x \%less\% y
}
\arguments{
\item{x, y}{vectors of the same mode containing sequences of items}
}
\value{
a vector of the same mode containing the difference of \code{x} less \code{y}
}
\description{
The set difference of x less y
}
\seealso{
\code{\link[spida2]{\%and\%}} and \code{\link[spida2]{\%or\%}}
}
