% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/less.R
\name{\%or\%}
\alias{\%or\%}
\title{union as a binary operator}
\usage{
x \%or\% y
}
\arguments{
\item{x, y}{vectors of the same mode containing sequences of items}
}
\value{
a vector of the same mode containing the union of \code{x} less \code{y}
}
\description{
The union of x and y
}
\seealso{
\code{\link[spida2]{\%and\%}} and \code{\link[spida2]{\%less\%}}
}
