% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{\%T\%}
\alias{\%T\%}
\title{Pipe for magrittr that passes LHS invisibly}
\usage{
lhs \%T\% rhs
}
\description{
Pipe for magrittr that passes LHS invisibly
}
\examples{
data(hs)
hs \%T\% dim  \%T\% names  \%>\% head 
}
