% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepv.R
\name{grepv}
\alias{grepv}
\title{Grep with value = T}
\usage{
grepv(...)
}
\arguments{
\item{\dots}{arguments to \code{\link{grep}}}
}
\description{
Equivalent of grep(..., value = TRUE)
}
\seealso{
\code{\link{grepl}}
}
