% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepv.R
\name{grepv2}
\alias{grepv2}
\title{Grep with value = T}
\usage{
grepv2(...)
}
\arguments{
\item{\dots}{arguments to \code{\link{grep}}}
}
\description{
Equivalent of grep(..., value = TRUE)
}
\details{
Now obsolete since the addition of the same command to base R.
}
\seealso{
\code{\link{grepl}}
}
