% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{gspf.1}
\alias{gspf.1}
\title{Marked for deletion}
\usage{
gspf.1(x, knots, degree = 3, smooth = pmax(pmin(degree[-1],
  degree[-length(degree)]) - 1, 0), intercept = 0, signif = 3)
}
\arguments{
\item{x}{FIXME}

\item{knots}{FIXME}

\item{degree}{FIXME}

\item{smooth}{FIXME}

\item{intercept}{FIXME}

\item{signif}{FIXME}
}
\description{
Marked for deletion
}
