% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_array.R
\name{kable_array}
\alias{kable_array}
\title{Print array using kableExtra}
\usage{
kable_array(a, caption = "", ...)
}
\arguments{
\item{a}{array}

\item{caption}{an optional caption for the printed table}

\item{...}{functions from kable extra and their arguments}
}
\value{
hmtl code to include in a 'asis' chunk
}
\description{
Print a higher-dimensional array with subheadings for the levels
defined by the third and higher dimensions of the array.
}
\examples{
kable_array(Titanic)
kable_array(Titanic, row_spec = list(0, angle = -30), 
  add_header_above =list(c(' '= 1,'Gender'=2,Freq = 1)),
  column_spec(1, bold = T, border_right = T)) 
\dontrun{  
kable_array(Titanic, row_spec = list(0, angle = -30), 
  add_header_above =list(c(' '= 1,'Gender'=2,Freq = 1)),
  column_spec(1, bold = T, border_right = T)) \%>\%
  save_kable(file = 'test.html', self_contained = T)
}
}
