% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R
\name{ker}
\alias{ker}
\title{Kernel (null space) of a linear transformation}
\usage{
ker(L, tol = 1e-14)
}
\arguments{
\item{L}{a matrix}

\item{tol}{the smallest singular value that is considered to
indicate a non-zero dimension.}
}
\description{
Kernel (null space) of a linear transformation
}
