% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{Lag}
\alias{Lag}
\title{Contextual lag with intrapolation}
\usage{
Lag(
  x,
  id = rep(1, length(x)),
  idx = 1:length(x),
  lag = 1,
  at = NULL,
  check = T
)
}
\arguments{
\item{x}{vector to be lagged}

\item{id}{clustering variable}

\item{idx}{time variable}

\item{lag}{(default 1)}

\item{at}{alternative to 'lag': evaluate x at a particular value of 'idx'}

\item{check}{(default TRUE) check for unique values of idx within each cluster}
}
\value{
return the value of 'x' in the position 'idx - lag' in the same cluster
}
\description{
Lags within clusters based on a time vector with possibly unequal intervals
}
\seealso{
\code{\link[spida2]{LagI}}, \code{\link[spida2]{DiffI}}, \code{\link[spida2]{Diff}},
}
