% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{LagI}
\alias{LagI}
\title{Contextual lag with intrapolation}
\usage{
LagI(x, id, time, lag = 1, delta = 0.01, check = T)
}
\arguments{
\item{x}{vector to be lagged}

\item{id}{clustering variable}

\item{lag}{(default 1)}

\item{delta}{(default .01)}

\item{check}{(default TRUE) check for unique values of idx within each cluster}

\item{idx}{time variable}
}
\value{
return the value of 'x' in the position 'idx - lag' in the same cluster
        calculated by intrapolation.
}
\description{
Lags within clusters based on a time vector with possibly unequal intervals
}
\seealso{
\code{\link[spida2]{LagI}}, \code{\link[spida2]{DiffI}}, \code{\link[spida2]{Diff}},
}
