% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Lall}
\alias{Lall}
\title{Hypothesis matrix to test for lmer objects}
\usage{
Lall(fit, nam)

Lall(fit, nam)
}
\arguments{
\item{fit}{model}

\item{pat}{FIXME}
}
\value{
hypothesis matrix

hypothesis matrix
}
\description{
Hypothesis matrix to test for lmer objects

Hypothesis matrix to test for lmer objects
}
