% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Lc}
\alias{Lc}
\title{Hypothesis matrix for lmer objects: comparisons with reference level}
\usage{
Lc(fit, nam, ref = 1, verbose = 0)

Lc(fit, nam, ref = 1, verbose = 0)
}
\arguments{
\item{fit}{model}

\item{nam}{name of categorical factor}

\item{ref}{FIXME}

\item{verbose}{default 0}
}
\description{
Hypothesis matrix for lmer objects: comparisons with reference level

Hypothesis matrix for lmer objects: comparisons with reference level
}
