% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Lcall}
\alias{Lcall}
\title{Construct hypothesis matrix to test ????}
\usage{
Lcall(fit, factors = getFactorNames(fit), debug = F)

Lcall(fit, factors = getFactorNames(fit), debug = F)
}
\arguments{
\item{fit}{model}

\item{factors}{FIXME}
}
\value{
a hypothesis matrix

a hypothesis matrix
}
\description{
Construct hypothesis matrix to test ????

Construct hypothesis matrix to test ????
}
