% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Ldiff.old}
\alias{Ldiff.old}
\title{Older version of Ldiff}
\usage{
Ldiff.old(
  fit,
  pat,
  levnames = c(reflevel, substring(rownames(L), cut + 1)),
  reflevel = "<ref>",
  cut = nchar(pat)
)

Ldiff.old(
  fit,
  pat,
  levnames = c(reflevel, substring(rownames(L), cut + 1)),
  reflevel = "<ref>",
  cut = nchar(pat)
)
}
\arguments{
\item{fit}{model}

\item{pat}{pattern to match in names of coefficients}

\item{levnames}{level names}

\item{reflevel}{name for reference level}

\item{cut}{number of characters in pattern to retain for labelling}
}
\value{
hypothesis matrix

hypothesis matrix
}
\description{
Older version of Ldiff

Older version of Ldiff
}
