% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff.rdc}
\alias{Ldiff.rdc}
\title{Version of Ldiff used in RDC}
\usage{
Ldiff.rdc(fit, nam, ref = "no longer used")
}
\arguments{
\item{fit}{model}

\item{nam}{name of effect}

\item{ref}{unused}
}
\value{
hypothesis matrix
}
\description{
Version of Ldiff used in RDC
}
