% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Lequal}
\alias{Lequal}
\title{Hypothesis matrix to test equality of factor level effects}
\usage{
Lequal(fit, pat)

Lequal(fit, pat)
}
\arguments{
\item{fit}{model}

\item{pat}{FIXME}
}
\value{
hypothesis matrix

hypothesis matrix
}
\description{
Hypothesis matrix to test equality of factor level effects

Hypothesis matrix to test equality of factor level effects
}
