% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels.R
\name{Levels}
\alias{Levels}
\title{Levels of a factor or unique values of a vector}
\usage{
Levels(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{(default TRUE) remove NAs if x is not a factor, otherwise just return levels(x)}
}
\description{
Useful as an argument to lapply
}
\examples{
\dontrun{
head(hs)
fits <- lapply(Levels(hs$school), function(sid)
           lm(mathach ~ ses, subset(hs, school == sid)))
coefs <- lapply(fits, coefs)
df <- do.call(rbind, coefs)
df$school <- names(df)
}
}

