% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{lin_comb}
\alias{lin_comb}
\title{Linear combination of x}
\usage{
lin_comb(x, w)
}
\arguments{
\item{x}{numerical vector}

\item{w}{vector of weights replicated to have same length as 'x'}
}
\description{
Linear combination of x
}
\seealso{
Other Post-stratification survey functions: 
\code{\link{jk_wtd_mean_se}()},
\code{\link{jk_wtd_means}()},
\code{\link{wtd_mean}()}
}
\concept{Post-stratification survey functions}
