% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llapply.R
\name{llapply}
\alias{llapply}
\title{Version of lapply that turns the input list into a list with elements of length one}
\usage{
llapply(x, FUN = function(x) x, ...)
}
\arguments{
\item{x}{a vector (atomic or list each of whose elements
will be made into a list of length one and inserted
into a list to be passed to \code{\link{lapply}}.}

\item{FUN}{a function to be applied to each element
of the list formed from \code{x}. If \code{FUN} is missing,
it defaults to the identity function, thus returning
x turned into a list of elements of length one.}

\item{\dots}{additional variables passed to \code{\link{lapply}}
if \code{FUN} returns a vector of length 1.}
}
\description{
Allows lapply to work on elements that are themselves
lists of length one so that the name of each element
is available within the function called by \code{\link{lapply}}.
}
