% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lmat}
\alias{Lmat}
\title{Generate a hypothesis matrix}
\usage{
Lmat(fit, pattern, fixed = FALSE, invert = FALSE, debug = FALSE)
}
\arguments{
\item{fit}{a fitted object with a \code{\link{getFix}} method.}

\item{pattern}{a regular expression that matches names of coefficients}
}
\value{
hypothesis matrix to test the hypothesis that the true values of
matched coefficients are zero
}
\description{
Generates a hypothesis matrix to test whether a group of coefficients in a
linear model are jointly zero, selecting the coefficients that match a
regular expression.
}
