% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lmu}
\alias{Lmu}
\title{Estimate predicted response for a factor level.}
\usage{
Lmu(fit, nam, verbose = 0)
}
\arguments{
\item{fit}{model}

\item{nam}{name of categorical effect}

\item{verbose}{default 0}
}
\description{
Estimate predicted response for a factor level.
}
