% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lrm}
\alias{Lrm}
\title{Construct hypothesis matrix to test repeated measures factor effects}
\usage{
Lrm(fit, nam, vals = 1:nrow(L.mu))
}
\value{
hypothesis matrix
}
\description{
Construct hypothesis matrix to test repeated measures factor effects
}
