% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{lssvd}
\alias{lssvd}
\alias{lssvd_old}
\title{lsfit using SVD}
\usage{
lssvd(x, y, zero = 1e-07, ...)

lssvd_old(x, y, zero = 10^(-16), has_intercept = all(cbind(x)[, 1] == 1))
}
\arguments{
\item{x}{a matrix of predictors. Should include an intercept term if needed}

\item{y}{a vector or matrix of responses}

\item{zero}{a value used to identify latent roots eseentially 0}

\item{has_intercept}{not yet used}
}
\value{
A list with three elements: Beta, the matrix of coefficients; residuals and sse
}
\description{
not well optimized yet
}
\section{Functions}{
\itemize{
\item \code{lssvd_old()}: previous version not using MASS::ginv

}}
