% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lfx.R
\name{M}
\alias{M}
\alias{M.formula}
\alias{M.factor}
\alias{M.M}
\alias{M.default}
\alias{*.M}
\alias{<.factor}
\alias{>.factor}
\alias{<=.factor}
\alias{>=.factor}
\title{Method to generate hypothesis matrices from formulas}
\usage{
M(x, ...)

\method{M}{formula}(form, ..., keep.intercept = FALSE)

\method{M}{factor}(x, base = NULL)

\method{M}{M}(x, ...)

\method{M}{default}(x, ...)

\method{*}{M}(e1, e2)

\method{<}{factor}(x, y)

\method{>}{factor}(x, y)

\method{<=}{factor}(x, y)

\method{>=}{factor}(x, y)
}
\arguments{
\item{...}{other arguments}

\item{form}{formula}

\item{keep.intercept}{(default: FALSE) if TRUE generate a column for the intercept term}
}
\description{
Interprets a formula to generate the corresponding portion of a hypothesis matrix. See examples in \code{\link[spida2]{Lfx}}. Creates an 'M' object with methods to generate coefficients to estimate interaction terms and comparisons of factor levels.
}
\section{Methods (by class)}{
\itemize{
\item \code{M(formula)}: formula method

\item \code{M(factor)}: factor method

\item \code{M(M)}: method for M object

\item \code{M(default)}: default method

}}
\section{Methods (by generic)}{
\itemize{
\item \code{ * }: multiplication of M objects

}}
\section{Functions}{
\itemize{
\item \code{ < }: '<' method for factors

\item \code{ > }: '>' method for factors

\item \code{ <= }: '<=' method for factors

\item \code{ >= }: '>=' method for factors

}}
\seealso{
\code{\link[spida2]{Lfx}}
}
