% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{make.grid}
\alias{make.grid}
\title{Create a data frame of predictors to display a model}
\usage{
make.grid(mod, n = 50, dmax = 5)

make.grid(mod, n = 50, dmax = 5)
}
\arguments{
\item{mod}{a model from which a model data frame can be
extracted with \code{\link{getModelData}}.}

\item{n}{number of values generated for continuous predictors. Default: 50}

\item{dmax}{maximum number of distinct values for a numeric variable
to be considered discrete. Default: 5}
}
\value{
a data frame with every combination of the values generated
for each variable.

a data frame with every combination of the values generated
for each variable.
}
\description{
Generates the cartesian product of factors and ranges
of continuous variables in a model.

Generates the cartesian product of factors and ranges
of continuous variables in a model.
}
\details{
This function is particularly useful to display fitted  values
of a model that is non-linear in continuous predictors where
plotting curves against the original data would produce
linear segments instead of a continuous curve.

By default, 50 values of each continuous variable is
generated. See the examples to generate fewer values.

Variables are considered discrete if the have fewer than
discrete_max distinct values.

This function is particularly useful to display fitted  values
of a model that is non-linear in continuous predictors where
plotting curves against the original data would produce
linear segments instead of a continuous curve.

By default, 50 values of each continuous variable is
generated. See the examples to generate fewer values.

Variables are considered discrete if the have fewer than
discrete_max distinct values.
}
