% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegrid.R
\name{makegrid}
\alias{makegrid}
\title{Cartesian product of model predictors}
\usage{
makegrid(fit, ...)
}
\arguments{
\item{fit}{a model for which \code{getD} returns the
model frame, assuming that the response is
in the first column}

\item{...}{predictors in the model for which a different
set of values will be used.}
}
\description{
Cartesian product of levels of predictor values in a model with
the option to specify desired values for some terms of the
model
}
\examples{
fit <- lm(mathach ~ ses * Sector, hs)
dd <- makegrid(fit, ses = seq(-2,2,.1))
library(car)
fit <- lm(income ~ type * education, Prestige)
dd <- makegrid(fit, type = 'prof', education = 6:18)
# note that dd$type will have the correct form, 
# i.e. a factor with the same three levels as those in 
# the original data set.
levels(dd$type)
       
}
