% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up.R
\name{na.include}
\alias{na.include}
\title{na.include action}
\usage{
na.include(obj)
}
\arguments{
\item{obj}{a data.frame whose factors will be redefined so exclude = NULL}
}
\description{
From the Hmisc package, author: Frank Harrell
}

