% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{name}
\alias{name}
\title{Modify the name of an object and return the renamed object}
\usage{
name(x, nam = x, ...)
}
\arguments{
\item{x}{object to named with a 'names' attribute}

\item{nam}{vector of names or a function}

\item{...}{additional arguments to used by \code{nam} if it is a function}
}
\value{
x with a modified name. If \code{nam} is a function then the names attribute
        is modified with \code{names(x) <- nam(names(x), ...)}
}
\description{
This function is well suited to modifying the name of an object in a pipeline.
It can be thought of as the verb corresponding to the noun 'names'.
}
\examples{
\dontrun{
x <- as.list(letters[1:3])
x
name(x)
library(magrittr)
x \%>\% name \%>\% name(toupper)
x \%>\% name \%>\% name(sub_, '(.*)', 'time_\\\\1')
}
}
