% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R
\name{ob}
\alias{ob}
\title{SVD to generate an othonormal basis}
\usage{
ob(x, tol = 1e-14)
}
\arguments{
\item{x}{a matrix}

\item{tol}{the smallest singular value that is considered to
indicate a non-zero dimension.}
}
\value{
an orthonormal basis for the column span of x
}
\description{
SVD to generate an othonormal basis
}
