% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paik.R
\name{paik}
\alias{paik}
\title{Paik-Agresti diagrams
New version of Paik-Agresti diagrams}
\usage{
paik(
  form,
  data,
  ylab,
  xlab = xn,
  zlab = zn,
  cex.factor = 10,
  cex.zlab = 1,
  ...
)
}
\arguments{
\item{form}{formula of the form y ~ x + z + count}

\item{data}{data frame}

\item{ylab}{optional parameter}

\item{xlab}{optional parameter}

\item{cex.factor}{default 10: scales the size of circles}

\item{...}{additional parameters for \code{lattice::xyplot} such
\code{xlim}, \code{ylim}, \code{main}, etc.}
}
\description{
Paik-Agresti diagrams
New version of Paik-Agresti diagrams
}
\examples{
paik(verdict ~ d.race + v.race | count, death.penalty)
paik(verdict ~ d.race + v.race | count, death.penalty, 
    ylab = 'verdict', main = 'Paik-Agresti Diagram')
paik(Status ~ Gender + Dept | count, Berkeley)
gd(lty = 1:4)
paik(Status ~ Gender + Dept | count, Berkeley,
    zlab = 'Department')
}
