% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pals}
\alias{pals}
\title{Display selected colors n at a time}
\usage{
pals(pattern = "", ignore = "^gray[0-9]|^grey[0-9]", n = 30, ask = FALSE)
}
\arguments{
\item{pattern}{match to select colors}

\item{ignore}{match to ignore colors (default 'grey|gray')}

\item{n}{maximum number to show in one page (default 30)}

\item{ask}{(default FALSE) prompt for new page}
}
\description{
Display selected colors n at a time
}
