% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch.R
\name{pch}
\alias{pch}
\alias{lty}
\title{Easy pch characters}
\usage{
pch(shape = 1, fill = 1)

lty(line = 1)
}
\arguments{
\item{shape}{'square', 'circle', 'triangle', 'diamond', 'ltriangle' (or 1, 2, 3, 4, 5 respectively). Default: 1}

\item{fill}{'empty', 'solid', 'fill' (with different color) (or 1, 2, 3 respectively). Default: 1}

\item{line}{'solid','dashed','dotted','dash dot','long dash','long dash dot' (or 1, 2, 3, 4, 5, 6 respectively). Default: 1}
}
\value{
value(s) of pch or lty to produce corresponding character or line type
}
\description{
Fifteen pch symbols and six line types can be generated
mnemonic names.
}
\section{Functions}{
\itemize{
\item \code{lty}: Easy line types
}}

