% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdConstruct.pdInd}
\alias{pdConstruct.pdInd}
\title{Construct pdInd object}
\usage{
pdConstruct.pdInd(
  object,
  value = numeric(0),
  form = formula(object),
  nam = Names(object),
  data = sys.parent(),
  cov = NULL,
  ...
)
}
\arguments{
\item{object}{an object inheriting from the class \code{pdInd}, representing
a positive definite matrix with zero covariances except in the first row and
column.}

\item{value}{an optional initialization value}

\item{form}{an optional one-sided linear formula specifying the row/column
names for the matrix represented by \code{object}.}

\item{nam}{and optional vector of character strings specifying the
row/column names for the matrix represented by \code{object}.}

\item{data}{and optional data frame i which to evaluate the variables names
in \code{value} and \code{form}. ...}

\item{cov}{optional position in lower triangle of covariances that are
estimated and, thus, possibly non-zero. The default is that the covariances
in the first column are estimated and possibly non-zero.}
}
\description{
This function is an alternative constructor for the pdMat class 
associated with object and is mostly used internally in other functions. 
See the documentation on the principal constructor function 
\code{\link[spida2]{pdInd}}.
}
