% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdMatrix.pdInd}
\alias{pdMatrix.pdInd}
\title{pdMatrix method for pdInd objects}
\usage{
pdMatrix.pdInd(object, factor = FALSE)
}
\arguments{
\item{object}{a pdInd object}

\item{factor}{should an upper-triangular factor be returned of the variance
matrix}
}
\value{
a variance matrix or it upper-triangular factor.
}
\description{
This function is the pdMatrix method for pdInd objects
}
