% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{PolyShift}
\alias{PolyShift}
\title{Transformation of polynomial coefficients to change origin

Useful to equate polynomial coefficients for a periodic spline}
\usage{
PolyShift(a, n)
}
\arguments{
\item{a}{center for transformed polynomial}

\item{n}{number of coefficients}
}
\description{
If \code{coef} is a vector of coefficients for a polynomial of the
form \code{ coef[1] + coef[2]*x + coef[3]*x^2 + coef[4]*x^3 }, then
\code{coef_tr <- PolyShift(a, 4) \%*\% coef} is a vector of coefficients
for the same polynomial centered at \code{a}, i.e.
\code{ coef_tr[1] + coef_tr[2]*(x-a) + coef_tr[3]*(x-a)^2 + coef_tr[4]*(x-a)^3 },
}
\examples{
coefs <- c(3,2,4)
x <- 3
n <- length(coefs)
a <- 2
sum( coefs * x ^ (0:(n-1))) 
sum( PolyShift(a,n)\%*\%coefs * (x -a)^(0:(n-1)))
}
