% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R
\name{print.wald_lrt}
\alias{print.wald_lrt}
\title{Print method for wald objects}
\usage{
\method{print}{wald_lrt}(x, round = 6, pround = 5, ...)
}
\arguments{
\item{x}{wald object}

\item{round}{FIXME}

\item{pround}{FIXME}

\item{\dots}{FIXME}
}
\value{
called for  FIXME
}
\description{
Print method for wald objects
}
\note{
This is a note
}
\examples{
# coming soon FIXME
}
\seealso{
\code{\link{wald}}
}
\author{
GM
}
