% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{Proj.1}
\alias{Proj.1}
\title{Projection matrix}
\usage{
Proj.1(x)
}
\arguments{
\item{x}{}
}
\description{
Projection matrix into column space of a matrix. Works even if the matrix
does not have full column rank.
}
