% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{Proj.test}
\alias{Proj.test}
\title{Test accuracy of projection method.}
\usage{
Proj.test(x, fun = Proj)
}
\arguments{
\item{x}{}

\item{fun}{}
}
\description{
Test accuracy of projection method.
}
