% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbind.R
\name{Rbind}
\alias{Rbind}
\title{Stack data frames vertically}
\usage{
Rbind(...)
}
\arguments{
\item{\dots}{data frames or objects that can be coerced to data frames}
}
\description{
Stack data frames vertically on top of each other with columns identified by variable names. Matrices are coerced to data frames.
}
\examples{
zd1 <- data.frame(x = 1:3, y = 2, z = 3, .unique. = 'a')
zd2 <- data.frame(x = 1:4, y = 'y', w = 1, m = matrix(1:8, nrow = 4))
zm1 <- cbind(x = 1:3, y = 2, z = 3)
zm2 <- cbind(x = 1:4, y = 3, w = 1)
Rbind(zm1, zm2, zd1, zd1, zd2)
}

