% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_keep}
\alias{render_keep}
\title{Render with Rmarkdown and keep intermediate files}
\usage{
render_keep(input, ...)
}
\arguments{
\item{input}{file to be rendered}

\item{...}{other arguments passed to \code{\link{rmarkdown::render}}.}
}
\description{
Following https://github.com/rstudio/rmarkdown/issues/17
add the YAML: \code{knit: spida2::render_keep} to a .R Rmarkdown
document with output YAML \code{output: html_document}.
}
