% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{rowdiffs}
\alias{rowdiffs}
\title{Pairwise differences of rows of a matrix}
\usage{
rowdiffs(L)

rowdiffs(L)
}
\arguments{
\item{L}{a matrix}
}
\description{
Returns all pairwise difference of rows of a matrix with
rownames formed with the corresponding rownames of the original 
matrix. The main application is to estimate differences between
treatment levels given a linear hypothesis matrix, L, in which 
each row estimates a function of model parameter given
a level of a categorical variable. \code{rowdiffs} provides 
an analog to differentiation
(see \code{\link{Lfx}}) for categorical variables.

Returns all pairwise difference of rows of a matrix with
rownames formed with the corresponding rownames of the original 
matrix. The main application is to estimate differences between
treatment levels given a linear hypothesis matrix, L, in which 
each row estimates a function of model parameter given
a level of a categorical variable. \code{rowdiffs} provides 
an analog to differentiation
(see \code{\link{Lfx}}) for categorical variables.
}
