% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpfmt.R
\name{rpfmt}
\alias{rpfmt}
\alias{rpfmt.default}
\alias{rpfmt.wald}
\title{Displaying p-values with estimates}
\usage{
rpfmt(x, ...)

\method{rpfmt}{default}(x, digits = 3, pdigits = 5)

\method{rpfmt}{wald}(w, ...)
}
\arguments{
\item{x}{a n x 2 matrix or data frame}

\item{digits}{(default: 5) number of digits for rounding coefficient}

\item{pdigits}{(default: 5) number of digits for rounding p-values}
}
\description{
Takes a 2 column matrix or data frame in which the first column has estimated
values and the second has p-values and formats them into a single column
with p-values in parentheses
}
\details{
works on first element of a \code{\link{wald}} list
}
\examples{
rpfmt( cbind( 'estimate' = rnorm(7), "p-values" = 10^c(-1,-2,-3,-4,-5,-6,-7)))
}

