% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{sampler}
\alias{sampler}
\title{Show available characters, colours, etc.}
\usage{
sampler(n = 24, all = FALSE)
}
\arguments{
\item{n}{FIXME}

\item{all}{default: FALSE}
}
\description{
Show available characters, colours, etc.
}
\seealso{
\code{\link[lattice]{lattice::show.settings}}
}
\concept{lattice}
