% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp.R
\name{sc}
\alias{sc}
\title{Hypothesis matrix fpr gemeral regression splines}
\usage{
sc(sp, x, D = 0, type = 1)
}
\arguments{
\item{sp}{a spline function generated by \code{\link{gsp}}}

\item{x}{points where spline is tested}

\item{D}{(default 0) order of derivative to test, 0 is the value of the spline, 1 is the first derivative, etc.}

\item{type}{(default 1) if there is a discontinuity at a knot, type specifies whether to estimate the limit from the left (0), the limit from the right (1) or the saltus -- limit from the right minus the limit from left (2)}
}
\description{
This function helps to build hypothesis matrices for general splines. See examples
in \code{\link{gsp}}.
}
\seealso{
\code{\link{wald}} \code{\link{gsp}}
}
