% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setwd_here}
\alias{setwd_here}
\title{Set working directory to directory of active R script}
\usage{
setwd_here()
}
\description{
Set working directory to directory of active R script
if script is not being knitted
}
