% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Smoking3}
\alias{Smoking3}
\title{Smoking data by country and gender in approximately 2004}
\format{
A data set with 24 variables with self-descriptive names:
\describe{
\item{\code{country}}{Country}
\item{\code{iso3}}{iso 3-letter country abbreviation}
\item{\code{region}}{Region: 3-letter region designation}
\item{\code{HealthExpPC.Govt.exch}}{Health Expenditures per capita by government in US dollars using current currency exchange rates}
\item{\code{HealthExpPC.Tot.ppp}}{Total Health Expenditures per capita in US dollars using purchasing power parity exchange rates}
\item{\code{HealthExpPC.Govt.ppp}}{Health Expenditures per capita by government in US dollars using purchasing power parity exchange rates}
\item{\code{HealthExpPC.Tot.exch}}{HealthExpPC.Tot.exch}
\item{\code{total}}{same as \code{HealthExpPC.Tot.ppp}}
\item{\code{govt}}{same as \code{HealthExpPC.Govt.ppp}}
\item{\code{private}}{equal to \code{total - govt}}
\item{\code{sex}}{Sex}
\item{\code{lifeexp.Birth}}{lifeexp.Birth}
\item{\code{lifeexp.At60}}{lifeexp.At60}
\item{\code{smoking.tobacco.current}}{smoking.tobacco.current}
\item{\code{smoking.tobacco.daily}}{smoking.tobacco.daily}
\item{\code{smoking.cig.current}}{smoking.cig.current}
\item{\code{smoking.cig.daily}}{smoking.cig.daily}
\item{\code{Pop.Total}}{Total population}
\item{\code{Pop.MedAge}}{Median age}
\item{\code{Pop.pCntUnder15}}{percent of population under 15 years of age}
\item{\code{Pop.pCntOver60}}{Pop.pCntOver60}
\item{\code{Pop.pCntAnnGrowth}}{Pop.pCntAnnGrowth}
\item{\code{consumption.cigPC}}{consumption.cigPC}
\item{\code{hiv_prev15_49}}{hiv_prev15_49}
}

A data set with 24 variables with self-descriptive names:
\describe{
\item{\code{country}}{Country}
\item{\code{iso3}}{iso 3-letter country abbreviation}
\item{\code{region}}{Region: 3-letter region designation}
\item{\code{HealthExpPC.Govt.exch}}{Health Expenditures per capita by government in US dollars using current currency exchange rates}
\item{\code{HealthExpPC.Tot.ppp}}{Total Health Expenditures per capita in US dollars using purchasing power parity exchange rates}
\item{\code{HealthExpPC.Govt.ppp}}{Health Expenditures per capita by government in US dollars using purchasing power parity exchange rates}
\item{\code{HealthExpPC.Tot.exch}}{HealthExpPC.Tot.exch}
\item{\code{total}}{same as \code{HealthExpPC.Tot.ppp}}
\item{\code{govt}}{same as \code{HealthExpPC.Govt.ppp}}
\item{\code{private}}{equal to \code{total - govt}}
\item{\code{sex}}{Sex}
\item{\code{lifeexp.Birth}}{lifeexp.Birth}
\item{\code{lifeexp.At60}}{lifeexp.At60}
\item{\code{smoking.tobacco.current}}{smoking.tobacco.current}
\item{\code{smoking.tobacco.daily}}{smoking.tobacco.daily}
\item{\code{smoking.cig.current}}{smoking.cig.current}
\item{\code{smoking.cig.daily}}{smoking.cig.daily}
\item{\code{Pop.Total}}{Total population}
\item{\code{Pop.MedAge}}{Median age}
\item{\code{Pop.pCntUnder15}}{percent of population under 15 years of age}
\item{\code{Pop.pCntOver60}}{Pop.pCntOver60}
\item{\code{Pop.pCntAnnGrowth}}{Pop.pCntAnnGrowth}
\item{\code{consumption.cigPC}}{consumption.cigPC}
\item{\code{hiv_prev15_49}}{hiv_prev15_49}
}
}
\usage{
Smoking3

Smoking3
}
\description{
Smoking data by country and gender in approximately 2004

Smoking data by country and gender in approximately 2004
}
\examples{
if(!require("p3d")) remotes::install_github('gmonette/p3d')
library(p3d)
library(spida2) 
data(Smoking3)
dd <- Smoking3
head(dd)
dd$cig <- dd$consumption.cigPC
dd$Cigarettes <- dd$consumption.cigPC
dd$pop <- dd$Pop.Total
dd$Life <- dd$lifeexp.Birth
dd$region <- factor(dd$region)
dd$log.total <- log(dd$total)
dd$Health <- dd$total

ds <- subset( dd, sex == "BTSX")  # Both sexes combined
rownames(ds) <- ds$country
Init3d(cex = 2)
Plot3d(Life ~ Cigarettes + Health | region, ds, phi = 0, theta = 0, fov = 0)
Plot3d(lifeexp.At60 ~ govt + private | region, ds, phi = 0, theta = 0, fov = 0)
Id3d()
Plot3d_par(col = c('#550000','#005500','#000055'))
Plot3d(lifeexp.At60 ~ govt + private | region, ds, phi = 0, theta = 0, fov = 0)

if(!require("p3d")) remotes::install_github('gmonette/p3d')
library(p3d)
library(spida2) 
data(Smoking3)
dd <- Smoking3
head(dd)
dd$cig <- dd$consumption.cigPC
dd$Cigarettes <- dd$consumption.cigPC
dd$pop <- dd$Pop.Total
dd$Life <- dd$lifeexp.Birth
dd$region <- factor(dd$region)
dd$log.total <- log(dd$total)
dd$Health <- dd$total

ds <- subset( dd, sex == "BTSX")  # Both sexes combined
rownames(ds) <- ds$country
Init3d(cex = 2)
Plot3d(Life ~ Cigarettes + Health | region, ds, phi = 0, theta = 0, fov = 0)
Plot3d(lifeexp.At60 ~ govt + private | region, ds, phi = 0, theta = 0, fov = 0)
Id3d()
Plot3d_par(col = c('#550000','#005500','#000055'))
Plot3d(lifeexp.At60 ~ govt + private | region, ds, phi = 0, theta = 0, fov = 0)

}
\keyword{datasets}
