% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{smspline.v}
\alias{smspline.v}
\title{smspline.v from library splines with better ginverse}
\usage{
smspline.v(time)
}
\arguments{
\item{time}{}
}
\description{
smspline.v from library splines with better ginverse
}
