% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{solve.pdInd}
\alias{solve.pdInd}
\title{solve method for pdInd objects.}
\usage{
\method{solve}{pdInd}(a, b, ...)
}
\arguments{
\item{a}{the pdInd object to invert.}

\item{b}{is unused but copied from \code{solve.pdLogChol}.}
}
\value{
a pdInd object corresponding to the matrix inverse of \code{a}.
}
\description{
This produces a pdInd object corresponding to the inverse of its argument.
}
