% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sortdf}
\alias{sortdf}
\title{Order the rows of a data frame}
\usage{
sortdf(data, form)
}
\arguments{
\item{data}{a data frame}

\item{form}{a formula identifying variables to be used in sorting}
}
\value{
The formula is evaluated using \code{model.frame} and the
result is used as the argument of \code{order} which, in turn is used to
order the data frame. The ordered data frame is returned.
}
\description{
Order the rows of a data frame using variables identified in a formula
evaluated in the data frame. Convenient in a pipeline.
}
\examples{
\dontrun{
require(car)
require(magrittr)
Prestige  \%>\% sortdf(~type+income) -> Prestige.ordered
}
}
