% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{spline.E}
\alias{spline.E}
\title{Spline estimation function}
\usage{
spline.E(knots, degree, smooth, lin = NULL, intercept = 0,
  signif = 3)
}
\arguments{
\item{knots}{}

\item{degree}{}

\item{smooth}{}

\item{intercept}{}

\item{signif}{}
}
\description{
Spline estimation function
}
