% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sub_}
\alias{sub_}
\alias{sub_.default}
\alias{sub_.factor}
\title{Methods providing versions of sub and gsub for pipelines and factors}
\usage{
sub_(x, pattern, replacement, ...)

\method{sub_}{default}(x, pattern, replacement, ...)

\method{sub_}{factor}(x, pattern, replacement, ...)
}
\arguments{
\item{x}{object to change. If x is a factor, the substitution is performed on its levels attribute.}

\item{pattern}{the 'from' regular expression}

\item{replacement}{the 'to' regular expression}

\item{...}{other arguments for  \code{\link{sub}} or  \code{\link{gsub}}}
}
\value{
Value of \code{sub(pattern, replacement, x, ...)} or \code{gsub(pattern, replacement, x, ...)}
}
\description{
The first argument of sub_  and gsub_ is the object to 
be modified instead of the pattern to be matched. Thus \code{sub_}
and \code{gsub_} can be used as generic functions that dispatch on
the class of the first argument. This allows them to recognize factors
and work with the levels attribute instead of coercing the
factor to a character vector. Two consequences are that a
factor, instead of a character vector, is returned, and that the
original order of the levels is preserved unless the number of 
levels is reduced.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{factor}: method for factor objects
}}

\examples{
\dontrun{
library(magrittr)
x <- as.list(1:4)
x  \%>\% name(letters[1:4])
x  \%>\% name(letters[1:4]) \%>\% name(toupper)
x  \%>\% name(letters[1:4]) \%>\% 
       name(sub_, '(.*)', 'Time_\\\\1')  \%>\% unlist
}
}
\seealso{
\code{\link{sub}} \code{\link{sub_}}
}
