% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sublast}
\alias{sublast}
\title{Substitute last occurrence of a pattern in a string}
\usage{
sublast(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{character string containing a regular expression whose
whose last match in 'x' will be replaced}

\item{replacement}{character string that is a replacement for the
matched 'pattern' in 'x'}

\item{x}{a character vector where matches are sought.}

\item{...}{arguments passed to \code{\link{gsub}}.}
}
\description{
Substitute last occurrence of a pattern in a string
}
\examples{
string <- 'a_b_c_D'
sublast('_','__',string)
dd <- data.frame(id = 1:3, X_a_1 = 1:3, X_a_2 = 1:3, X_b_1 = 1:3, X_b_2 = 1:3)
dd
names(dd) <- sublast('_','__',names(dd))
tolong(dd, sep = '__')
tolong(dd, sep = '__') \%>\% tolong(sep = '_', idvar = 'id2', timevar = 'time2')

}
