% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{jk_lin_comb_se}
\alias{std}
\alias{survey}
\title{Functions for post-stratification adjustments in surveys}
\usage{
std(x, div = sum(x))

jk_lin_comb_se(x, by, w, check = TRUE)
}
\arguments{
\item{x}{vector of weights}

\item{div}{divisor (default: sum(x))}

\item{by}{stratification variable or list of stratification variable. Default: single level for unstratified samples}

\item{w}{vector of weights, e.g. Horvitz-Thomson weights or renormed version. Default: 1}
}
\description{
Functions for post-stratification adjustments in surveys

### Functions for linear combinations

Post-stratification Jacknife estimate of SE for a linear combination
}
\details{
Standardize a vector of weights
}
\seealso{
\code{\link{wtd_mean}} for weighted means, \code{link{lin_comb}} for linear
     combinations, \code{\link{jk_wtd_mean_se}} for a jacknife estimate of the SE of a weighted mean
     and \code{\link{jk_lin_comb_se}} for a jacknife estimate of the SE of a linear combination.
}

