% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab__}
\alias{tab__}
\alias{Tab}
\alias{pab}
\title{tab without marginal totals and 'All'}
\usage{
tab__(..., names_to_drop = c("All", "Total"))
}
\arguments{
\item{\dots}{arguments to the \code{\link{tab}} function.}

\item{names_to_drop}{(default "Total") names of margins to drop}
}
\description{
Version of \code{\link{tab}}, with the option to drop selected margins without
necessarily dropping
the marginal average proportions denoted by "All".
}
\seealso{
\code{\link{tab}}
}
