% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabf}
\alias{tabf}
\title{Create an array of data frames splitting on a formula with marginal faces}
\usage{
tabf(dd, fmla, all = TRUE, ...)
}
\arguments{
\item{dd}{data frame}

\item{fmla}{formula whose RHS creates a table and
LHS adds variables to the resulting data frames}

\item{all}{adds all variables in the data frame}
}
\description{
Like a table but each element consists of the rows of the data frame
for a particular combination of levels of the formula
}
\examples{
tabf(mtcars, ~ cyl + carb)[1,1]            
tabf(mtcars, ~ cyl + carb) \%>\% dimnames  
tabf(mtcars, ~ cyl + carb + gear ) \%>\% Apply(dim) \%>\% Apply(paste, collapse = ' ')  
          
}
