% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{test1}
\alias{test1}
\alias{test2}
\title{test1 to see how to combine different function}
\usage{
test1(x, y)

test2(x, y, z)
}
\arguments{
\item{x}{the x parameter of test2}

\item{y}{the y parameter of test2}

\item{z}{the z parameter of test2}
}
\description{
test1 second paragraph

test2 second paragraph -- goes in description
}
\details{
test1 third paragraph

test2 third paragraph -- goes in details
}
\section{test1}{

special info for test1
}

\section{test2}{

special info for test2
}

