% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Unemp}
\alias{Unemp}
\title{Monthly unemployment data in the U.S. from 1995-01-01 to 2019-02-01}
\format{
A data set with 290 rows an two variables:
  \describe{
    \item{\code{date}}{from 1995-01-01 to 2019-02-01}
    \item{\code{unemployment}}{rate in the U.S.}
  }
}
\usage{
Unemp
}
\description{
Monthly unemployment data in the U.S. from 1995-01-01 to 2019-02-01
}
\keyword{datasets}
