% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up.R
\name{up_apply}
\alias{up_apply}
\title{Apply a function to clusters of rows in a data frame}
\usage{
up_apply(object, form, FUN, ..., sep = "/")
}
\arguments{
\item{object}{a data frame as source for an aggregated result}

\item{form}{a one-sided formula identifying the variable(s) in \code{object}
that identifies clusters. e.g. ~ school/Sex to get a summary within each Sex
of each school}

\item{FUN}{a function to be applied to each data frame consisting of a
cluster of rows of 'object'. The most common choice is \code{\link{with}} so that
'...' can be an expression using variable names in 'object'.}

\item{...}{other arguments to FUN, frequently when FUN is 'with', an expression
using variable names in 'object'}
}
\description{
Apply a function to clusters of rows in a data frame and return
the result so it is conformable with the data frame created by
'up' applied to the same data frame and same clustering formula
}
\examples{
zd <- data.frame(a=c('a','a','b','b','c','c','c'),
      b = c("B","B","A","B","C","D","D"), x = 1:7, y = 11:17)
zd$n <- capply(zd$x, zd[c('a','b')], length)
zdu <- up(zd, ~a, agg = ~b)
zdu
zdu$p <- up_apply(zd, ~ a, with, sum(x)/sum(y))
zdu
}
