% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up.R
\name{varLevel}
\alias{varLevel}
\title{Identify level of aggregation at which a variable in invariant}
\usage{
varLevel(x, form, ...)
}
\arguments{
\item{form}{formula to identify clusters}

\item{\dots}{extra arguments to \code{\link{constant}} function}
}
\value{
Level 0 is a constant for the whole data frame,
Level <= 1 implies the variable is constant within levels of id1,
Level <= 2 implies the variable is constant within levels of id2,
... etc.

NOTE: NA counts as a distinct value
}
\description{
Shows levels of each variable with respect to grouping formula
of form ~id or nested ids ~id1/id2
}
