% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Vcov}
\alias{Vcov}
\alias{Vcor}
\title{Generic 'vcov' extended to objects with a \code{getFix} method}
\usage{
Vcov(fit, ...)

Vcor(fit, ...)
}
\arguments{
\item{fit}{model with \code{\link{getFix}} method}
}
\value{
estimated variance covariance matrix of fixed effects
}
\description{
Generic 'vcov' extended to objects with a \code{getFix} method

Correlation matrix of fixed effects
}
\section{Functions}{
\itemize{
\item \code{Vcor}: correlation matrix of fixed effects
}}

\author{
GM
}
