% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{Weekdays}
\alias{Weekdays}
\title{Extract day of the week as a factor}
\usage{
Weekdays(date, first = "Sunday", ...)
}
\arguments{
\item{date}{an object that can be converted to a 'Date'}

\item{first}{day of the week, default: "Sunday"}
}
\description{
Extract day of the week as a factor
}
