% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{Xmat}
\alias{Xmat}
\title{Estimate Dth derivative of
polynomial at x}
\usage{
Xmat(x, degree, D = 0, signif = 3)
}
\arguments{
\item{x}{vector of values where derivative evaluated}

\item{degree}{of polynomial}

\item{signif}{significant digits for labeling}

\item{order}{of derivative}
}
\description{
Generate X matrix
to estimate Dth derivative
of polynomial at x
}
