% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{xqplot}
\alias{xqplot}
\alias{xqplot_}
\title{Extended Quantile Plots}
\usage{
xqplot(
  x,
  ptype = "quantile",
  labels = dimnames(x)[[2]],
  ...,
  mfrow = findmfrow(ncol(x)),
  ask = FALSE,
  mcex = 0.8,
  maxlab = 12,
  debug = isTRUE(options("debug")),
  mar = c(4, 2.5, 4, 1),
  xlab.pos = 2,
  xlab.cex = 0.7,
  sublab.pos = xlab.pos + xlab.cex * 1.15,
  text.cex.factor = 1,
  left.labs = F,
  class = TRUE,
  xaxs = "i",
  maxvarnamelength = 20
)

xqplot_(
  x,
  ptype = "quantile",
  labels = names(x),
  ...,
  mfrow = findmfrow(length(x)),
  ask = FALSE,
  mcex = 0.8,
  maxlab = 12,
  debug = F,
  mar = c(4, 2.5, 4, 1),
  xlab.pos = 2,
  xlab.cex = 0.7,
  sublab.pos = xlab.pos + xlab.cex * 1.15,
  text.cex.factor = 1,
  left.labs = F,
  class = TRUE,
  xaxs = "i",
  maxvarnamelength = 20
)
}
\arguments{
\item{x}{a data frame or list of variables to plot}

\item{ptype}{"quantile" (default) or "normal": kind of quantile to plot on x
axis.}

\item{labels}{names for each plot}

\item{\dots}{additional arguments passed to 'plot' command}

\item{mfrow}{number of rows and columns per page. If missing, an attempt is
made to choose a reasonable number.}

\item{ask}{if TRUE pause after each page, default: FALSE}

\item{mcex}{character expansion factor for marginal text
\code{mcex}}

\item{maxlab}{maximum number of categories to label in barcharts}

\item{debug}{if TRUE, print additional information}

\item{mar}{size of margins}

\item{xlab.pos}{position of xlab, default: 2}

\item{xlab.cex}{cex for xlab, default: .7}

\item{sublab.pos}{position for sublab, default: xlab.pos + xlab.cex * 1.15}

\item{text.cex.factor}{character expansion factor for barchart labels}

\item{left.labs}{determines placement of barchart labels}

\item{class}{show class of object, default: TRUE}

\item{xaxs}{style of x axis, default: 'i'}

\item{maxvarnamelength}{maximum length of variable name without splitting on
two lines.}
}
\description{
An easy way to see a dataset's variables at a glance. Shows uniform quantile
plot for numerical varibles and barcharts for factors. Quantile plots also
show a horizontal line at the position of the mean and at mean plus or minus
one standard deviation.
}
\section{Functions}{
\itemize{
\item \code{xqplot_()}: experimental version to handle data.frames with list elements

}}
\note{
Bugs:
'mfrow' should take the total number of variables into account if they will
fill more than one page so the last page is close to being full.

The current version of the function could be made much simpler and more
transparent. Some code is redundant.
}
\examples{
library(car)
library(spida2)
xqplot(Prestige)
xqplot(Prestige,"n") # normal quantiles
xqplot(Prestige, pch = '.', cex = 4, col = 'red')
# from the car package
qqPlot(Prestige$education)
qqPlot(Prestige$income)
# For exponential distribution use the 'distribution root'
# for the quantile function, e.g. 'exp' for 'qexp':
qqPlot(Prestige$income, distribution = 'exp')
qqPlot(Prestige$income, distribution = 'unif')

dd <- data.frame(x = 1:10, y = rnorm(10))
dd$listvar <- list(1,'A',2,4,5,6,7,8,9,10)
dd
xqplot_(dd)
}
