% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_misc.R
\name{yscale.components.log100real}
\alias{yscale.components.log100real}
\title{Plot '100 x log scale' data with real scale on right}
\usage{
yscale.components.log100real(...)
}
\description{
Plots axis on right side of lattice graph showing pretty selection of 
real values corresponding to 100 x log values in graph. The function has no 
parameters and is intended to be supplied as an argument to the 
`yscale.components` parameter in `xyplot``
}
\examples{
\dontrun{
library(car)
library(lattice)
log100 <- function(x) 100*log(x)
(ob1 <- xyplot(log100(income) ~ education | type, Prestige, groups = type))
(ob2 <- update(ob1, ylab = expression(plain(log)[e](income) \%*\% 100)))
(ob3 <- update(ob2, 
              scales = list(y = list(alternating=3)), 
              ylab.right = 'income',
              yscale.components = 
                yscale.components.log100real
   )
)
}
}
\concept{
lattice
}
