% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapply.R
\name{zzaply}
\alias{zzaply}
\title{Double zap function to be documented in zapply}
\usage{
zzaply(x, ...)
}
\arguments{
\item{x}{vector in zzapply}
}
\description{
Double zap function to be documented in zapply
}
