% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{acond}
\alias{acond}
\alias{aprop}
\title{Transform a frequency table into relative frequencies relative to a margin.}
\usage{
acond(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")

aprop(x, MARGIN = NULL, total.margins = TRUE, all.label = "All")
}
\arguments{
\item{x}{a table}

\item{MARGIN}{on which to condition}

\item{all.label}{label for marginal distribution}

\item{total.margin}{logical whether to include a total margin (default TRUE)}
}
\description{
Transform a frequency table into relative frequencies relative to a margin.
}

