% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brace.R
\name{brace}
\alias{brace}
\alias{brace_}
\title{Draw a brace}
\usage{
brace(x1 = 0, y1 = 0, x2 = 0, y2 = 1, right = TRUE, rad = 0.2,
  tip = FALSE)

brace_(...)
}
\arguments{
\item{x1, }{y1, x1, y2 endpoints of brace}

\item{right}{(default TRUE) direction in which brace points}

\item{rad}{(default 0.2) radius of curl of brace}

\item{tip}{the tip of the brace (default FALSE)}
}
\value{
the coordinates of the point of the brace, invisibly
}
\description{
Draw a brace
}
\section{Functions}{
\itemize{
\item \code{brace_}: default tip = TRUE
}}

\examples{
plot(c(0,10), c(0,10), type = 'n')
lines(brace(0, 0, 0, 5), lwd = 2)
text(brace(0, 0, 0, 5, tip = T),"size", adj = c(-.1,.3))
lines(brace(2, 0, 8, 5, right = F), lwd = 2)
text(brace_(2, 0, 8, 5, right = F), 'tip')
text(brace_(2, 0, 8, 5, right = F), 'tip', adj = c(.5,-.3), col = 'red')
plot(c(0,10),c(0,1))
brace(5,0,5,1) \%>\% lines
lapply(1:10, function(x) {
   lines(brace(x,0,x,.1*x,F))
   lines(brace(x,0,x-.2*x,.1*x,F),col = 'red')
})
}
\author{
Georges Monette, Michael Friendly
}
