% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{cols}
\alias{cols}
\title{Colour collection}
\usage{
cols()
}
\description{
A collection of colours found in various places
}
\examples{
names(cols())
pal(cols()$mycols)
pal(cols()$tol8)
lapply(names(cols()),function(x) pal(cols()[[x]], main = x))
}
