% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvar.R, R/dvar.R
\name{cvar}
\alias{cvar}
\alias{cvar.factor}
\alias{cvar.default}
\alias{dvar}
\title{Create a contextual variable for regression}
\usage{
cvar(x, id, all, na.rm, ...)

\method{cvar}{factor}(x, id, all = FALSE, na.rm = TRUE, ...)

\method{cvar}{default}(x, id, all, na.rm = TRUE, ...)

cvar(x, id, all, na.rm, ...)

\method{cvar}{factor}(x, id, all = FALSE, na.rm = TRUE, ...)

\method{cvar}{default}(x, id, all, na.rm = TRUE, ...)

dvar(x, id, all, na.rm, ...)
}
\arguments{
\item{x}{variable to be centered or residualized within groups. If x is a
factor, cvar and dvar return matrices whose columns are named consistently
with the names of coding variables for factors.}

\item{id}{identifies clusters}

\item{all}{(default FALSE) if TRUE cvar.factor returns the columns means of
an incidence matrix including the first level. Otherwise, the first level is
dropped for use in a linear model.}

\item{na.rm}{(default TRUE) whether to drop missing values}

\item{FUN}{(default mean) function to be applied to create contextual values}

\item{x}{variable to be centered or residualized within groups. If x is a
factor, cvar and dvar return matrices whose columns are named consistently
with the names of coding variables for factors.}

\item{id}{identifies clusters}

\item{all}{(default FALSE) if TRUE cvar.factor returns the columns means of
an incidence matrix including the first level. Otherwise, the first level is
dropped for use in a linear model.}

\item{na.rm}{(default TRUE) whether to drop missing values}

\item{x}{variable to be centered or residualized within groups. If x is a
factor, cvar and dvar return matrices whose columns are named consistently
with the names of coding variables for factors.}

\item{id}{identifies clusters}

\item{all}{(default FALSE) if TRUE cvar.factor returns the columns means of
an incidence matrix including the first level. Otherwise, the first level is
dropped for use in a linear model.}

\item{na.rm}{(default TRUE) whether to drop missing values}
}
\description{
cvar and  See also \code{\link{dvar}} are 
designed to be used in regression formulas to
create a contextual mean of a cluster-varying variable and
a 'centered-within-groups' version.

cvar and  See also \code{\link{dvar}} are 
designed to be used in regression formulas to
create a contextual mean of a cluster-varying variable and
a 'centered-within-groups' version.

cvar and dvar are designed to be used in regression formulas to
create a contextual mean of a cluster-varying variable and
a 'centered-within-groups' version. See also \code{\link{cvar}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: method for class 'factor'

\item \code{default}: default method

\item \code{factor}: method for class 'factor'

\item \code{default}: default method
}}

\examples{
\dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
}
\dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
}
\dontrun{
dd <- data.frame(x= 1:100, id = rep( LETTERS[1:10], each = 10))
dd$a <- factor(sample( c('a','b','c'), 100, replace = T))
dd$y <- dd$x + rep(rnorm(10), each = 10) + rnorm(100) + as.numeric(dd$a)
library(nlme)
fit <- lme( y ~ x + cvar(x,id), dd, random = ~ 1 + dvar(x,id) | id)
anova( fit , type = 'm')
# The output of 'anova' can be used to test whether a contextual variable
# should be included in the model
}
}
