% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropone.R
\name{dropone}
\alias{dropone}
\alias{parDropone}
\title{Drop each row or cluster from a fitted object}
\usage{
dropone(
  fit,
  form = NULL,
  FUN = if (inherits(fit, "lme")) fixef else coef,
  data = getData(fit),
  ...
)

parDropone(
  fit,
  form = NULL,
  FUN = if (inherits(fit, "lme")) fixef else coef,
  data = getData(fit),
  mc.cores = detectCores(),
  ...
)
}
\arguments{
\item{fit}{a fitted object with an \code{\link{update}} method
and a \code{\link{getData}} method.}

\item{form}{a formula evaluated in the data frame for 'fit'
that defines clusters to be dropped one at a time. If NULL, the
default, each row is dropped successively.}

\item{FUN}{function used to extract coefficients. Default:
\code{\link[spida2]{fixef}} for class 'lme' and
\code{\link{coef}} otherwise.}

\item{data}{data frame in which to evaluate the refitted object. Default:
'getData(fit)'.}

\item{mc.cores}{number of cores to use. Default: \code{parallel::detectCores()} \eqn{\exp^n}}
}
\description{
Drop each row or each cluster in a fitted object and
collect the resulting linear estimates. Return
the drop one estimates along with variable from original
data that are invariant within clusters, as well as the maximum
value of DFBETAS for each cluster dropped.
}
\section{Functions}{
\itemize{
\item \code{parDropone()}: parallel computing version of dropone using the parallel package

}}
