% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{getFactorNames}
\alias{getFactorNames}
\title{get the names of variables that are factors}
\usage{
getFactorNames(object, ...)
}
\arguments{
\item{object}{a data frame or object with a \code{\link{getData}} method that produces a data frame or a list}

\item{...}{other orguments (currently not used)}
}
\description{
get the names of variables that are factors
}

