% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{getFix}
\alias{getFix}
\alias{getFix.multinom}
\alias{getFix.lm}
\alias{getFix.glm}
\alias{getFix.lme}
\alias{getFix.gls}
\alias{getFix.lmer}
\alias{getFix.glmerMod}
\alias{getFix.glmer}
\alias{getFix.mer}
\alias{getFix.zeroinfl}
\alias{getFix.mipo}
\alias{getFix.MCMCglmm}
\alias{getFix.stanfit}
\alias{getFix.lmerMod}
\alias{getFix.default}
\title{Get information on fixed effects from a model object}
\usage{
getFix(fit, ...)

\method{getFix}{multinom}(fit, ...)

\method{getFix}{lm}(fit, robust = FALSE, type = "HC3", ...)

\method{getFix}{glm}(fit, robust = FALSE, ...)

\method{getFix}{lme}(fit, robust = FALSE, ...)

\method{getFix}{gls}(fit, robust = FALSE, ...)

\method{getFix}{lmer}(fit, robust = FALSE, ...)

\method{getFix}{glmerMod}(fit, robust = FALSE, ...)

\method{getFix}{glmer}(fit, robust = FALSE, ...)

\method{getFix}{mer}(fit, robust = FALSE, ...)

\method{getFix}{zeroinfl}(fit, robust = FALSE...)

\method{getFix}{mipo}(fit, robust = FALSE, ...)

\method{getFix}{MCMCglmm}(fit, robust = FALSE, ...)

\method{getFix}{stanfit}(fit, pars, include = TRUE, robust = FALSE, ...)

\method{getFix}{lmerMod}(fit, robust = FALSE, ...)

\method{getFix}{glmerMod}(fit, robust = FALSE, ...)

\method{getFix}{default}(fit, ...)
}
\arguments{
\item{fit}{A fitted model object}

\item{\dots}{Other arguments \link{unused}}
}
\value{
Returns a list with the following components:
\itemize{
\item{fixed}{Fixed effect parameter estimates}
\item{vcov}{Covariance matrix of the parameters}
\item{df}{denominator degrees of freedom for each effect}
}
}
\description{
\code{getFix} extracts coefficients, covariance matrix and degrees of
freedom from model objects. Its main purpose is to extract information need
by the \code{wald} function. To extend the wald function to a new class of
objects, it is merely necessary to write a method for \code{getFix}.
}
\details{
Extending the \code{\link{wald}} function to a new class of objects only
requires a \code{getFix} method for the new class.
}
\section{Methods (by class)}{
\itemize{
\item \code{getFix(multinom)}: method for multinom objects in package nnet

\item \code{getFix(lm)}: method for lm objects

\item \code{getFix(glm)}: method for glm objects

\item \code{getFix(lme)}: method for lme objects in the nlme package

\item \code{getFix(gls)}: method for gls objects in the nlme package

\item \code{getFix(lmer)}: method for lmer objects in the lme4 package

\item \code{getFix(glmerMod)}: method for glmerMod objects in the lme4 package

\item \code{getFix(glmer)}: method for glmer objects in the lme4 package

\item \code{getFix(mer)}: method for mer objects in the lme4 package

\item \code{getFix(zeroinfl)}: method for zeroinfl objects in the pscl?? package

\item \code{getFix(mipo)}: method for mipo objects in the mice package

\item \code{getFix(MCMCglmm)}: method for MCMCglmm objects in the MCMCglmm package

\item \code{getFix(stanfit)}: method for \code{stanfit} objects in the \code{rstan} package

\item \code{getFix(lmerMod)}: method for \code{lmerMod} objects in the \code{lme4} package

\item \code{getFix(glmerMod)}: method for \code{glmerMod} objects in the \code{lme4} package

\item \code{getFix(default)}: print message if getFix id used for a class for which a method has not been written

}}
\examples{
library(nlme)
fit <- lme( mathach ~ (ses + I(ses^2)) * Sex, hs, random = ~ 1 + ses| school)
getFix(fit)

data(Prestige, package="car")
mod.prestige <- lm(prestige ~ education + income, data=Prestige)
getFix(mod.prestige)

}
\seealso{
\code{\link{wald}}
}
\author{
Georges Monette
}
