% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{here}
\alias{here}
\title{Directory or filename of active R script}
\usage{
here(dir = FALSE)
}
\arguments{
\item{dir}{(default: TRUE) return the directory of the current R script, otherwise the full file name}
}
\description{
Uses the \code{this.path} package return the file name
of the current R script or its enclosing directory.
}
