% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{labs}
\alias{labs}
\alias{labs.default}
\alias{labs.data.frame.lab}
\alias{labs<-}
\alias{print.data.frame.lab}
\alias{[.data.frame.lab}
\alias{labs<-.data.frame}
\alias{labs<-.default}
\title{Assign, extract and print labels}
\usage{
labs(x, ...)

\method{labs}{default}(x, ...)

\method{labs}{data.frame.lab}(x, ...)

labs(x, ...) <- value

\method{print}{data.frame.lab}(x, ..., digits = NULL, quote = FALSE,
  right = TRUE)

\method{[}{data.frame.lab}(x, ...)

\method{labs}{data.frame}(x) <- value

\method{labs}{default}(x) <- value
}
\arguments{
\item{x}{an object}

\item{\dots}{additional arguments (currently not used)}

\item{digits}{to print}

\item{quote}{(default FALSE) put quotes around strings}

\item{right}{(default TRUE) justfication}
}
\description{
Assigns, extracts and prints labels for various objects
}
