% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Ldiff}
\alias{Ldiff}
\title{Hypothesis matrix to test differencs in factor levels}
\usage{
Ldiff(fit, pat, levnames = c(reflevel, substring(rownames(L), cut + 1)),
  reflevel = "<ref>", cut = nchar(pat), verbose = F)
}
\value{
hypothesis matrix
}
\description{
Hypothesis matrix to test differencs in factor levels
}
