% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{Lequal}
\alias{Lequal}
\title{Hypothesis matrix to test equality of factor level effects}
\usage{
Lequal(fit, pat)
}
\value{
hypothesis matrix
}
\description{
Hypothesis matrix to test equality of factor level effects
}

