% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{Lrm}
\alias{Lrm}
\title{Construct hypothesis matrix to test repeated measures factor effects}
\usage{
Lrm(fit, nam, vals = 1:nrow(L.mu))

Lrm(fit, nam, vals = 1:nrow(L.mu))
}
\arguments{
\item{fit}{model}

\item{nam}{name of categorical effect}

\item{vals}{select rows of mean estimation matrix FIXME}
}
\value{
hypothesis matrix

hypothesis matrix
}
\description{
Construct hypothesis matrix to test repeated measures factor effects

Construct hypothesis matrix to test repeated measures factor effects
}
