% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran.R
\name{map}
\alias{map}
\title{Apply a mapping to a vector}
\usage{
map(from, to, x)
}
\description{
A mnemonic for \code{to[match(x, from)]}. Compare with the more elaborate \code{\link[yscs]{tran}}.
}
\examples{
x.from <- c('A',"FromNA",'Z','B',"N")
x.to   <- factor(c('a',NA,'z','b',NA),levels=c('z','a','b',NA))
x <- c("Z","B",NA,"N","M")
map(x.from, x.to, x)
}

