% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{name}
\alias{name}
\alias{col_name}
\alias{row_name}
\title{Modify the name of an object and return the renamed object}
\usage{
name(x, nam = x, ...)

col_name(x, nam = x, ...)

row_name(x, nam = x, ...)
}
\arguments{
\item{x}{object to be named or renamed with a 'names', 'colnames'
or 'rownames' attribute}

\item{nam}{vector of names or a function}

\item{...}{additional arguments to used by \code{nam} if it is a function}
}
\value{
x with a modified name. If \code{nam} is a function then the names attribute
is modified with \code{names(x) <- nam(names(x), ...)}.
}
\description{
This function is well suited to modifying the name of an object in a pipeline.
It can be thought of as the verb corresponding to the noun 'names'.
Convenient functions to modify names using regular expressions are
\code{\link{spida2::sub_}} and \code{\link{spida2::gsub_}}.
See also \code{\link{uname}}.
}
\examples{
\dontrun{
x <- as.list(letters[1:3])
x
name(x)
library(magrittr)
x \%>\% name \%>\% name(toupper)
x \%>\% name \%>\% name(sub_, '(.*)', 'time_\\\\1')
}
}
