% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{name}
\alias{name}
\title{Modify the name of an object and return the renamed object}
\usage{
name(x, nam = x, ...)
}
\arguments{
\item{x}{object}

\item{nam}{name}
}
\value{
x with a modified name
}
\description{
This function is well suited to modifying the name of an object in a pipeline.
}
\examples{
\dontrun{
x <- as.list(letters[1:3])
x
library(magrittr)
x  \%>\% name
gsub_ <- function(x,pat,to,...) gsub(pat,to,x,...)
x \%>\% name()
x  \%>\% name \%>\% name(toupper)
x  \%>\% name \%>\% name(sub_, '(.*)', 'time_\\\\1')
}
}
