% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConjComp.R
\name{OrthoComp}
\alias{OrthoComp}
\title{Orthogonal basis for the orthogonal complement of the column space of a matrix.}
\usage{
OrthoComp(X, Z, tol = 1e-07)
}
\description{
Orthogonal basis for the orthogonal complement of the column space of a matrix.
}

