% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paik.R
\name{paik}
\alias{paik}
\title{Paik-Agresti diagrams}
\usage{
paik(
  formula,
  counts,
  resp.lvl = 2,
  data,
  circle.mult = 0.4,
  xlab = NULL,
  ylab = NULL,
  leg.title = NULL,
  leg.loc = NULL,
  show.mname = FALSE,
  pal = "Dark2",
  col = rep(brewer.pal(length(cl), pal), length(cl)),
  ...
)
}
\arguments{
\item{formula}{A two sided formula, e.g. Y ~ X1 + X2, with
cross-classified categorical variables. The second
explanatory variable, i.e. X2, is used as the trace variable
whose levels are distinguished in the graph with different
colors. Interactions and nested terms are not allowed.}

\item{counts}{A vector of counts for the associated
categorical variables in formula.}

\item{resp.lvl}{The level in Y of primary interest. See
example below.}

\item{data}{Dataframe containing variables in formula.}

\item{circle.mult}{Multiplier for circle radii in the diagram.}

\item{xlab}{X-axis label. By default this is defined as the categories
in the first explanatory variable, X1.}

\item{ylab}{Y-axis label. By default these will be proportions with
respect to the specified level of interest in the response.}

\item{leg.title}{Legend title. By default the conditioning variable name.}

\item{leg.loc}{Legend location. A legend location keyword; "bottomright",
"bottom", "bottomleft", "left", "topleft", "top",
"topright", "right" or "center".}

\item{show.mname}{Logical, indicating whether or not the words "Marginal prop"
should printed in the graph above the dotted line indicating
marginal proportions.}

\item{pal}{which palette to use from the RColorBrewer package, default: 'Dark2'}

\item{col}{colors for groups, default brewer.pal(n, pal) where n is the number of levels of the conditioning variables}

\item{...}{Additional arguments from plot.}
}
\description{
Construct a Paik-Agresti diagram adapted and enhanced from the
\code{asbio::paik} function.
}
\examples{
require(tcltk)

data(death.penalty)# from Agresti 2012 

op <- par(mfrow=c(1,2), mar=c(4,4,0,0))
paik(verdict ~ d.race + v.race, counts = count, data = death.penalty, 
leg.title = "Victims race", xlab = "Defendants race", 
ylab = "Proportion receiving death penalty")
par(mar=c(4,2,0,2))
paik(verdict ~ v.race + d.race, counts = count, data = death.penalty, 
xlab = "Victims race", leg.title = "Defendants race",leg.loc="topleft", 
ylab = "", yaxt = "n")
par(op)

}
\references{
Agresti, A. (2012) Categorical Data Analysis, 3rd edition.
  New York. Wiley.
  Paik M. (1985) A graphical representation of a three-way
  contingency table: Simpson's paradox and correlation.
  American Statistician 39:53-54.
}
\author{
Ken Aho
}
