% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdInd}
\alias{pdInd}
\title{Constuctor for Positive-Definite Matrix With Zero Covariances Between
Predictor Random Effects}
\usage{
pdInd(value = numeric(0), form = NULL, nam = NULL, data = sys.parent(),
  cov = NULL, zero = NULL)
}
\arguments{
\item{value}{an optional initialization value, which can be any of the
following ...}

\item{form}{an optional one-sided linear formula specifying the row/column
names for the matrix represented by \code{object}.}

\item{nam}{and optional vector of character strings specifying the
row/column names for the matrix represented by \code{object}.}

\item{data}{and optional data frame i which to evaluate the variables names
in \code{value} and \code{form}. ...}

\item{cov}{optional position in lower triangle of covariances that are
estimated and, thus, possibly non-zero. The default is that the covariances
in the first column are estimated and possibly non-zero.}

\item{zero}{optional way of specifying covariances constrained to be equal
to zero. Only the lower triangular portion of the \code{zero} is used. The
elements that are equal to 0 corresponds to the pattern of elements that are
constrained to zero in the covariance matrix.}
}
\description{
This function is a constructor for the \code{pdInd} class, representing a
positive-definite matrix with zero covariances except possibly in the first
row and column. If the matrix associated with \code{object} is of dimension
$n$, it is represented by $n + (n-1)$ unrestricted parameters representing a
lower-triangular log-Cholesky decomposition. The first $n$ parameters are
the logs of the diagonal elements of the matrix and the last $n-1$
components are the $n-1$ remaining elements of the lower-triangular
decomposition corresponding the to the possibly non-zero covariances in the
first row.
}
\section{CAUTION}{

cov and zero do not work. Until fixed, \code{pdInd} only creates the
default covariance pattern in which the only non-zero covariances are those with the first element.
}

