% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp_util.R
\name{PolyShift}
\alias{PolyShift}
\title{Transformation of polynomial coefficients to change origin
 
Useful to equate polynomial coefficients for a periodic spline}
\usage{
PolyShift(a, n)
}
\arguments{
\item{a}{center for transformed polynomial}

\item{n}{number of coefficients}
}
\description{

}
\examples{
coefs <- c(3,2,4)
x <- 3
n <- length(coefs)
a <- 2
sum( coefs * x ^ (0:(n-1))) 
sum( PolyShift(a,n)\%*\%coefs * (x -a)^(0:(n-1)))
}
