% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-lrt.R, R/wald.R
\name{print.cat}
\alias{print.cat}
\title{Print method for 'cat' objects}
\usage{
\method{print}{cat}(object, ...)

\method{print}{cat}(object, ...)
}
\arguments{
\item{\dots}{unused arguments}

\item{x}{a cat object to be printed with \code{\link{cat}}}
}
\value{
invisible(x)

invisible(x)
}
\description{
Print method for 'cat' objects

Print method for 'cat' objects
}
