% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab_}
\alias{Tab}
\alias{pab}
\alias{tab_}
\title{tab without marginal totals}
\usage{
tab_(..., names_to_drop = "Total")
}
\arguments{
\item{names_to_drop}{(default "Total") names of margins to drop}

\item{\dots}{arguments to the \code{\link{tab}} function.}
}
\description{
Version of \code{\link{tab}}, with the option to drop selected margins without
necessarily dropping
the marginal average proportions denoted by "All".
}
\seealso{
\code{\link{tab}}
}

