% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{xqplot}
\alias{xqplot}
\title{Extended Quantile Plots}
\usage{
xqplot(x, ptype = "quantile", labels = dimnames(x)[[2]], ...,
  mfrow = findmfrow(ncol(x)), ask = prod(mfrow) < ncol(x) &&
  dev.interactive(), mcex = 0.8, maxlab = 12, debug = F, mar = c(5, 2,
  3, 1), text.cex.factor = 1, left.labs = F, maxvarnamelength = 20)
}
\arguments{
\item{x}{a data frame or list of variables to plot}

\item{ptype}{"quantile" (default) or "normal": kind of quantile to plot on x
axis.}

\item{labels}{names for each plot}

\item{\dots}{additional arguments passed to 'plot' command}

\item{mfrow}{number of rows and columns per page. If missing, an attempt is
made to choose a reasonable number.}

\item{mcex}{character expansion factor for marginal text
\code{mcex}}

\item{maxlab}{maximum number of categories to label in barcharts}

\item{debug}{if TRUE, print additional information}

\item{mar}{size of margins}

\item{text.cex.factor}{character expansion factor for barchart labels}

\item{left.labs}{determines placement of barchart labels}

\item{maxvarnamelength}{maximum length of variable name without splitting on
two lines.}
}
\description{
An easy way to see a dataset's variables at a glance. Shows uniform quantile
plot for numerical varibles and barcharts for factors. Quantile plots also
show a horizontal line at the position of the mean and at mean plus or minus
one standard deviation.
}
\note{
Bugs:
'mfrow' should take the total number of variables into account if they will
fill more than one page so the last page is close to being full.

The current version of the function could be made much simpler and more
transparent. Some code is redundant.
}
\examples{
require(car)
xqplot(Prestige)
xqplot(Prestige,"n") # normal quantiles
}
