% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{years}
\alias{years}
\alias{years.Date}
\title{Get the year of a Date object}
\usage{
years(x, ...)

\method{years}{Date}(x, ...)
}
\arguments{
\item{x}{a Date object}
}
\value{
the year as an integer
}
\description{
Works like 'months', 'weekdays', etc. for years.
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: method for Date objects
}}

